Further separate http/socket routes
