+++ app.js
@@ -2,7 +2,6 @@
 const express = require('express')
 const session = require('express-session') // For storing client login data
 const { encrypt, decrypt } = require('./crypto.js') // For encrypting passwords
-const sqlite3 = require('sqlite3').verbose()
 const jwt = require('jsonwebtoken') // For authentication system between Plugins and Formbar
 const excelToJson = require('convert-excel-to-json')
 const multer = require('multer') // Used to upload files
@@ -12,8 +11,10 @@ const fs = require("fs")
 const { isAuthenticated, isLoggedIn, permCheck } = require('./modules/authentication.js')
 const { logger } = require('./modules/logger.js')
 const { logNumbers, settings } = require('./modules/config.js')
-const { MANAGER_PERMISSIONS, TEACHER_PERMISSIONS, GUEST_PERMISSIONS, STUDENT_PERMISSIONS, MOD_PERMISSIONS, BANNED_PERMISSIONS, DEFAULT_CLASS_PERMISSIONS } = require('./modules/permissions.js')
-const { classInformation } = require('./modules/class.js')
+const { MANAGER_PERMISSIONS, TEACHER_PERMISSIONS, GUEST_PERMISSIONS, STUDENT_PERMISSIONS, MOD_PERMISSIONS, BANNED_PERMISSIONS, DEFAULT_CLASS_PERMISSIONS, CLASS_SOCKET_PERMISSIONS, CLASS_SOCKET_PERMISSION_SETTINGS, GLOBAL_SOCKET_PERMISSIONS } = require('./modules/permissions.js')
+const { classInformation, Classroom } = require('./modules/class.js')
+const { Student } = require('./modules/student.js')
+const database = require('./modules/database.js')
 
 const app = express()
 const http = require('http').createServer(app)
@@ -22,7 +23,6 @@ const io = require('socket.io')(http)
 // Set EJS as our view engine
 app.set('view engine', 'ejs')
 
-
 // Create session for user information to be transferred from page to page
 const sessionMiddleware = session({
 	secret: crypto.randomBytes(256).toString('hex'), //Used to sign into the session via cookies
@@ -30,11 +30,9 @@ const sessionMiddleware = session({
 	saveUninitialized: false //Forces a session that is new, but not modified, or "uninitialized" to be saved to the session store
 })
 
-
 // Sets up middleware for the server by calling sessionMiddleware
 // adds session middleware to express
 app.use(sessionMiddleware)
-
 
 // For further uses on this use this link: https://socket.io/how-to/use-with-express-session
 // Uses a middleware function to successfully transmit data between the user and server
@@ -46,7 +44,6 @@ io.use((socket, next) => {
 // Allows express to parse requests
 app.use(express.urlencoded({ extended: true }))
 
-
 // Use a static folder for web page assets
 app.use(express.static(__dirname + '/static'))
 app.use('/js/chart.js', express.static(__dirname + '/node_modules/chart.js/dist/chart.umd.js'))
@@ -54,14 +51,11 @@ app.use('/js/iro.js', express.static(__dirname + '/node_modules/@jaames/iro/dist
 app.use('/js/floating-ui-core.js', express.static(__dirname + '/node_modules/@floating-ui/core/dist/floating-ui.core.umd.min.js'))
 app.use('/js/floating-ui-dom.js', express.static(__dirname + '/node_modules/@floating-ui/dom/dist/floating-ui.dom.umd.min.js'))
 
-// Establishes the connection to the database file
-const db = new sqlite3.Database('database/database.db')
-
 /*This line is executing a SQL query using the get method from a db object. The get method is typically part of a SQLite database interface in Node.js.
 The SQL query is SELECT MAX(id) FROM poll_history, which retrieves the maximum is value from the poll_history table. This is typically the id of the
 most recently added record. The result of the query is passed to a callback function as the second argument (pollHistory), and any error that occurred
 during the execution of the query is passed as the first argument (err).*/
-db.get('SELECT MAX(id) FROM poll_history', (err, pollHistory) => {
+database.get('SELECT MAX(id) FROM poll_history', (err, pollHistory) => {
 	/*This is an error handling block. If an error occurred during the execution of the SQL query (i.e., if err is not null), then the error is logged
 	using a logger object's log method. The log method is called with two arguments: a string indicating the severity level of the log ('error'), and 
 	the stack trace of the error (err.stack).*/
@@ -103,143 +97,6 @@ const PASSIVE_SOCKETS = [
 	'classBannedUsersUpdate'
 ]
 
-/*This line declares a constant object named GLOBAL_SOCKET_PERMISSIONS. The const keyword mean that the variable cannot be reassigned. However,
-the properties of the object can still be modified.*/
-const GLOBAL_SOCKET_PERMISSIONS = {
-	//This represents an event that changes permissions which requires manager permissions.
-	permChange: MANAGER_PERMISSIONS,
-	//This represents an event that deletes a class which requires teacher permissions.
-	deleteClass: TEACHER_PERMISSIONS,
-	//This represents an event that gets all classes you own which requires teacher permissions.
-	getOwnedClasses: TEACHER_PERMISSIONS,
-	//This represents an event that logs you out which requires guest permissions.
-	logout: GUEST_PERMISSIONS,
-	//This represents an event that gets all classes you are in which requires guest permissions.
-	getUserClass: GUEST_PERMISSIONS,
-	//This represents an event that deletes a user which requires manager permissions.
-	deleteUser: MANAGER_PERMISSIONS,
-	//This represents an event that updates the manager panel which requires manager permissions.
-	managerUpdate: MANAGER_PERMISSIONS,
-	//This represents an event that updates the ip address list which requires manager permissions.
-	ipUpdate: MANAGER_PERMISSIONS,
-	//This represents an event that adds an ip address to the list which requires manager permissions.
-	addIp: MANAGER_PERMISSIONS,
-	//This represents an event that removes an ip address from the list which requires manager permissions.
-	removeIp: MANAGER_PERMISSIONS,
-	//This represents an event that changes an ip address on the list which requires manager permissions.
-	changeIp: MANAGER_PERMISSIONS,
-	//This represents an event that gets the ip address list which requires manager permissions.
-	toggleIpList: MANAGER_PERMISSIONS,
-	//This represents an event that saves the tags list which requires teacher permissions.
-	saveTags: TEACHER_PERMISSIONS,
-	//This represents an event that creates a new tag which requires teacher permissions.
-	newTag: TEACHER_PERMISSIONS,
-	//This represents an event that removes a tag which requires teacher permissions.
-	removeTag: TEACHER_PERMISSIONS,
-	//This represents an event that submits a password change request which requires student permissions.
-	passwordRequest: STUDENT_PERMISSIONS,
-	//This represents an event that approves the password change request which requires manager permissions.
-	approvePasswordChange: MANAGER_PERMISSIONS,
-	//This represents an event that updates the password which requires manager permissions.
-	passwordUpdate: MANAGER_PERMISSIONS
-}
-
-//This line declares a constant object CLASS_SOCKET_PERMISSIONS. The const keyword means that the variable can't be reassigned.
-const CLASS_SOCKET_PERMISSIONS = {
-	//This line defines a property named help that requires student permissions, which was defined earlier in the code. 
-	help: STUDENT_PERMISSIONS,
-	/*These lines define actions like responding to a poll, requesting a break, ending a break, updating a poll, updating the mode, updating a quiz,
-	and updating a lesson. All of these actions require student permissions.*/
-	pollResp: STUDENT_PERMISSIONS,
-	requestBreak: STUDENT_PERMISSIONS,
-	endBreak: STUDENT_PERMISSIONS,
-	pollUpdate: STUDENT_PERMISSIONS,
-	modeUpdate: STUDENT_PERMISSIONS,
-	quizUpdate: STUDENT_PERMISSIONS,
-	lessonUpdate: STUDENT_PERMISSIONS,
-	//These lines define actions like updating the virtual bar, the virtual bar timer, and leaving a class. These actions require guest permissions.
-	vbUpdate: GUEST_PERMISSIONS,
-	vbTimer: GUEST_PERMISSIONS,
-	leaveClass: GUEST_PERMISSIONS,
-	//This line defines an action that updates the control panel, which requires mod permissions.
-	cpUpdate: MOD_PERMISSIONS,
-	//This line defines an action that displays a previous poll, which requires teacher permissions.
-	previousPollDisplay: TEACHER_PERMISSIONS,
-	//This line defines an action of updating the plugin list, which requires student permissions.
-	pluginUpdate: STUDENT_PERMISSIONS,
-	//This line defines an action of setting the class default permission setting, which requires manager permissions.
-	setClassPermissionSetting: MANAGER_PERMISSIONS,
-	//This line defines an action that starts a class poll, which requires mod permissions.
-	classPoll: MOD_PERMISSIONS,
-	//These lines define actions like creating a timer, and turning it on, which require teacher permissions.
-	timer: TEACHER_PERMISSIONS,
-	timerOn: TEACHER_PERMISSIONS
-}
-
-// make a better name for this
-const CLASS_SOCKET_PERMISSION_SETTINGS = {
-	/*This line maps the action startPoll to the permission associated with controlPolls. This means that in order to start a poll, a user must 
-	have the permissions associated with controlPolls.*/
-	startPoll: 'controlPolls',
-	//Similarly, to clear a poll, a user must have the permissions associated with controlPolls.
-	clearPoll: 'controlPolls',
-	//To end a poll, a user must have the permissions associated with controlPolls.
-	endPoll: 'controlPolls',
-	//To update the custom poll list, a user must have the permissions associated with controlPolls.
-	customPollUpdate: 'controlPolls',
-	//To save a poll to the custom poll list, a user must have the permissions associated with controlPolls.
-	savePoll: 'controlPolls',
-	//To delete a poll from the custom poll list, a user must have the permissions associated with controlPolls.
-	deletePoll: 'controlPolls',
-	//To set a poll as public, a user must have the permissions associated with controlPolls.
-	setPublicPoll: 'controlPolls',
-	//To share a poll with a user, a user must have the permissions associated with controlPolls.
-	sharePollToUser: 'controlPolls',
-	//To remove a shared poll with a user, a user must have the permissions associated with controlPolls.
-	removeUserPollShare: 'controlPolls',
-	//To get the ids of shared polls, a user must have the permissions associated with controlPolls.
-	getPollShareIds: 'controlPolls',
-	//To share a poll with a class, a user must have the permissions associated with controlPolls.
-	sharePollToClass: 'controlPolls',
-	//To remove a shared poll from a class, a user must have the permissions associated with controlPolls.
-	removeClassPollShare: 'controlPolls',
-	//To perform the next step in a lesson, a user must have the permissions associated with controlPolls.
-	doStep: 'controlPolls',
-	//To change permissions in a class, a user must have the permissions associated with manageStudents.
-	classPermChange: 'manageStudents',
-	//To kick a user from a class, a user must have the permissions associated with manageStudents.
-	classKickUser: 'manageStudents',
-	//To kick all users from a class, a user must have the permissions associated with manageStudents.
-	classKickStudents: 'manageStudents',
-	//To approve a break, a user must have the permissions associated with breakAndHelp.
-	approveBreak: 'breakAndHelp',
-	//To delete a ticket, a user must have the permissions associated with breakAndHelp.
-	deleteTicket: 'breakAndHelp',
-	//To change a plugin in the plugin list, a user must have the permissions associated with manageClass.
-	changePlugin: 'manageClass',
-	//To add a plugin to the plugin list, a user must have the permissions associated with manageClass.
-	addPlugin: 'manageClass',
-	//To remove a plugin from the plugin list, a user must have the permissions associated with manageClass.
-	removePlugin: 'manageClass',
-	//To end a class, a user must have the permissions associated with manageClass.
-	endClass: 'manageClass',
-	//To change the mode of a class, a user must have the permissions associated with manageClass.
-	modechange: 'manageClass',
-	//To update the list of banned users in a class, a user must have the permissions associated with manageStudents.
-	classBannedUsersUpdate: 'manageStudents',
-	//To ban a user from a class, a user must have the permissions associated with manageStudents.
-	classBanUser: 'manageStudents',
-	//To unban a user from a class, a user must have the permissions associated with manageStudents.
-	classUnbanUser: 'manageStudents',
-}
-
-// Variables
-let runningTimers = {};
-let currentPoll = 0
-let whitelistedIps = {}
-let blacklistedIps = {}
-
-
 // Add currentUser and permission constants to all pages
 app.use((req, res, next) => {
 	if (req.session.class)
@@ -257,101 +114,6 @@ app.use((req, res, next) => {
 
 	next()
 })
-
-// This class is used to create a student to be stored in the sessions data
-class Student {
-	// Needs username, id from the database, and if permissions established already pass the updated value
-	// These will need to be put into the constructor in order to allow the creation of the object
-	constructor(
-		username,
-		id,
-		permissions = STUDENT_PERMISSIONS,
-		API,
-		ownedPolls = [],
-		sharedPolls = [],
-		tags,
-		displayName
-	) {
-		this.username = username
-		this.id = id
-		this.permissions = permissions
-		this.classPermissions = null
-		this.tags = tags
-		this.ownedPolls = ownedPolls || []
-		this.sharedPolls = sharedPolls || []
-		this.pollRes = {
-			buttonRes: '',
-			textRes: '',
-			time: null
-		}
-		this.help = false
-		this.break = false
-		this.quizScore = ''
-		this.API = API
-		this.pogMeter = 0
-		this.displayName = displayName
-	}
-}
-
-
-// This class is used to add a new classroom to the session data
-// The classroom will be used to add lessons, do lessons, and for the teacher to operate them
-class Classroom {
-	// Needs the name of the class you want to create
-	constructor(id, className, key, permissions, sharedPolls, pollHistory, tags) {
-		this.id = id
-		this.className = className
-		this.students = {}
-		this.sharedPolls = sharedPolls || []
-		this.poll = {
-			status: false,
-			responses: {},
-			textRes: false,
-			prompt: '',
-			weight: 1,
-			blind: false,
-			requiredTags: [],
-			studentBoxes: [],
-			studentIndeterminate: [],
-			lastResponse: [],
-			allowedResponses: []
-		}
-		this.key = key
-		this.lesson = {}
-		this.activeLesson = false
-		this.steps
-		this.currentStep = 0
-		this.quiz = false
-		this.mode = 'poll'
-		this.permissions = permissions
-		this.pollHistory = pollHistory || []
-		this.tagNames = tags || [];
-		this.timer = {
-			startTime: 0,
-			timeLeft: 0,
-			active: false,
-			sound: false
-		}
-	}
-}
-
-//allows quizzes to be made
-class Quiz {
-	constructor(numOfQuestions, maxScore) {
-		this.questions = []
-		this.totalScore = maxScore
-		this.numOfQuestions = numOfQuestions
-		this.pointsPerQuestion = this.totalScore / numOfQuestions
-	}
-}
-
-//allows lessons to be made
-class Lesson {
-	constructor(date, content) {
-		this.date = date
-		this.content = content
-	}
-}
 
 // Functions
 // General functions
@@ -449,7 +211,7 @@ function joinClass(username, code) {
 			logger.log('info', `[joinClass] username=(${username}) classCode=(${code})`)
 
 			// Find the id of the class from the database
-			db.get('SELECT id FROM classroom WHERE key=?', [code], (err, classroom) => {
+			database.get('SELECT id FROM classroom WHERE key=?', [code], (err, classroom) => {
 				try {
 					if (err) {
 						reject(err)
@@ -464,7 +226,7 @@ function joinClass(username, code) {
 					}
 
 					// Find the id of the user who is trying to join the class
-					db.get('SELECT id FROM users WHERE username=?', [username], (err, user) => {
+					database.get('SELECT id FROM users WHERE username=?', [username], (err, user) => {
 						try {
 							if (err) {
 								reject(err)
@@ -478,7 +240,7 @@ function joinClass(username, code) {
 							}
 
 							// Add the two id's to the junction table to link the user and class
-							db.get('SELECT * FROM classusers WHERE classId=? AND studentId=?', [classroom.id, user.id], (err, classUser) => {
+							database.get('SELECT * FROM classusers WHERE classId=? AND studentId=?', [classroom.id, user.id], (err, classUser) => {
 								try {
 									if (err) {
 										reject(err)
@@ -507,7 +269,7 @@ function joinClass(username, code) {
 										logger.log('verbose', `[joinClass] cD=(${classInformation})`)
 										resolve(true)
 									} else {
-										db.run('INSERT INTO classusers(classId, studentId, permissions, digiPogs) VALUES(?, ?, ?, ?)',
+										database.run('INSERT INTO classusers(classId, studentId, permissions, digiPogs) VALUES(?, ?, ?, ?)',
 											[classroom.id, user.id, classInformation[code].permissions.userDefaults, 0], (err) => {
 												try {
 													if (err) {
@@ -558,7 +320,7 @@ function camelCaseToNormal(str) {
 
 function runQuery(query, params) {
 	return new Promise((resolve, reject) => {
-		db.run(query, params, (err) => {
+		database.run(query, params, (err) => {
 			if (err) reject(new Error(err))
 			else resolve()
 		})
@@ -567,7 +329,7 @@ function runQuery(query, params) {
 
 function getAll(query, params) {
 	return new Promise((resolve, reject) => {
-		db.all(query, params, (err, rows) => {
+		database.all(query, params, (err, rows) => {
 			if (err) reject(new Error(err))
 			else resolve(rows)
 		})
@@ -599,191 +361,29 @@ async function setClassOfApiSockets(api, classCode) {
 }
 
 async function getIpAccess(type) {
-	let ipList = await getAll(`SELECT id, ip FROM ip_${type}`)
+	const ipList = await getAll(`SELECT id, ip FROM ip_${type}`)
 	return ipList.reduce((ips, ip) => {
 		ips[ip.id] = ip
 		return ips
 	}, {})
 }
 
-// Socket.io functions
-async function managerUpdate() {
-	let [users, classrooms] = await Promise.all([
-		new Promise((resolve, reject) => {
-			db.all('SELECT id, username, permissions, displayName FROM users', (err, users) => {
-				if (err) reject(new Error(err))
-				else {
-					users = users.reduce((tempUsers, tempUser) => {
-						tempUsers[tempUser.username] = tempUser
-						return tempUsers
-					}, {})
-					resolve(users)
-				}
-			})
-		}),
-		new Promise((resolve, reject) => {
-			db.get('SELECT * FROM classroom', (err, classrooms) => {
-				if (err) reject(new Error(err))
-				else resolve(classrooms)
-			})
-		})
-	])
-
-	io.emit('managerUpdate', users, classrooms)
-}
-
-async function passwordRequest(newPassword, username) {
-	if (newPassword && username) {
-		let passwordChange = true;
-		io.emit("passwordUpdate", passwordChange, username, newPassword);
-	};
-};
-
-/**
-	 * Emits an event to sockets based on user permissions
-	 * @param {string} event - The event to emit
-	 * @param {string} classCode - The code of the class
-	 * @param {{permissions?: number, classPermissions?: number, api?: boolean, username?: string}} options - The options object
-	 * @param  {...any} data - Additional data to emit with the event
-	 */
-async function advancedEmitToClass(event, classCode, options, ...data) {
-	let classData = classInformation[classCode]
-
-	let sockets = await io.in(`class-${classCode}`).fetchSockets()
-
-	for (let socket of sockets) {
-		let user = classData.students[socket.request.session.username]
-		let hasAPI = false
-
-		if (!user) continue
-
-		if (options.permissions && user.permissions < options.permissions) continue
-		if (options.classPermissions && user.classPermissions < options.classPermissions) continue
-		if (options.username && user.username != options.username) continue
-
-		for (let room of socket.rooms) {
-			if (room.startsWith('api-')) {
-				hasAPI = true
-				break
-			}
-		}
-		if (options.api == true && !hasAPI) continue
-		if (options.api == false && hasAPI) continue
-
-		socket.emit(event, ...data)
-	}
-}
-
-
-//import routes
+// Import routes
 const apiRoutes = require('./routes/api.js')(classInformation)
+const routeFiles = fs.readdirSync('./routes/').filter(file => file.endsWith('.js'));
 
-//add routes to express
+for (const routeFile of routeFiles) {
+	// Skip for now as it's already handled
+	if (routeFile == "api.js") {
+		continue;
+	}
+
+	const route = require(`./routes/${routeFile}`);
+	route.run(app, io);
+}
+
+// Add routes to express
 app.use('/api', apiRoutes)
-
-// check if ip is banned
-app.use((req, res, next) => {
-	let ip = req.ip
-	if (ip.startsWith('::ffff:')) ip = ip.slice(7)
-
-	if (settings.whitelistActive && Object.keys(whitelistedIps).length > 0) {
-		const isWhitelisted = Object.values(whitelistedIps)
-			.some(value => ip.startsWith(value.ip))
-
-		if (!isWhitelisted) {
-			res.render('pages/message', {
-				message: 'Your IP has been banned',
-				title: 'Banned'
-			})
-			return
-		}
-	}
-	if (settings.blacklistActive && Object.keys(blacklistedIps).length > 0) {
-		const isBlacklisted = Object.values(blacklistedIps)
-			.some(value => ip.startsWith(value.ip))
-
-		if (isBlacklisted) {
-			res.render('pages/message', {
-				message: 'Your IP has been banned',
-				title: 'Banned'
-			})
-			return
-		}
-	}
-
-	next()
-})
-
-// This is the root page, it is where the users first get checked by the home page
-// It is used to redirect to the home page
-// This allows it to check if the user is logged in along with the home page
-// It also allows for redirection to any other page if needed
-app.get('/', isAuthenticated, (req, res) => {
-	try {
-		logger.log('info', `[get /] ip=(${req.ip}) session=(${JSON.stringify(req.session)})`)
-		if (classInformation[req.session.class].students[req.session.username].classPermissions >= TEACHER_PERMISSIONS) {
-			res.redirect('/controlPanel')
-		} else {
-			res.redirect('/student')
-		}
-	} catch (err) {
-		logger.log('error', err.stack);
-		res.render('pages/message', {
-			message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-			title: 'Error'
-		})
-	}
-})
-
-
-// A
-//The page displaying the API key used when handling oauth2 requests from outside programs such as formPix
-app.get('/apikey', isLoggedIn, (req, res) => {
-	try {
-		logger.log('info', `[get /apikey] ip=(${req.ip}) session=(${JSON.stringify(req.session)})`)
-
-		res.render('pages/apiKey', {
-			title: 'API Key',
-			API: classInformation[req.session.class].students[req.session.username].API
-		})
-	} catch (err) {
-		logger.log('error', err.stack);
-		res.render('pages/message', {
-			message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-			title: 'Error'
-		})
-	}
-})
-
-// B
-
-// C
-
-app.get('/changepassword', (req, res) => {
-	try {
-		res.render("pages/changepassword", {
-			title: "Change Password"
-		})
-	} catch (err) {
-		logger.log("error", err.stack);
-	}
-});
-
-app.post("/changepassword", (req, res) => {
-	try {
-		if (req.body.newPassword != req.body.confirmPassword) {
-			res.render("pages/message", {
-				message: "Passwords do not match",
-				title: "Error"
-			});
-		} else {
-			passwordRequest(req.body.newPassword, req.body.username);
-			res.redirect("/login");
-		}
-	} catch (err) {
-		logger.log("error", err.stack);
-	};
-});
 
 // An endpoint for the teacher to control the formbar
 // Used to update students permissions, handle polls and their corresponsing responses
@@ -994,7 +594,7 @@ app.post('/createClass', isLoggedIn, permCheck, (req, res) => {
 							permissions[permission] = DEFAULT_CLASS_PERMISSIONS[permission]
 						}
 					}
-					db.run('UPDATE classroom SET permissions=? WHERE key=?', [JSON.stringify(permissions), key], (err) => {
+					database.run('UPDATE classroom SET permissions=? WHERE key=?', [JSON.stringify(permissions), key], (err) => {
 						if (err) logger.log('error', err.stack)
 					})
 				}
@@ -1026,13 +626,13 @@ app.post('/createClass', isLoggedIn, permCheck, (req, res) => {
 			}
 
 			// Add classroom to the database
-			db.run('INSERT INTO classroom(name, owner, key, permissions, tags) VALUES(?, ?, ?, ?, ?)', [className, req.session.userId, key, JSON.stringify(DEFAULT_CLASS_PERMISSIONS), null], (err) => {
+			database.run('INSERT INTO classroom(name, owner, key, permissions, tags) VALUES(?, ?, ?, ?, ?)', [className, req.session.userId, key, JSON.stringify(DEFAULT_CLASS_PERMISSIONS), null], (err) => {
 				try {
 					if (err) throw err
 
 					logger.log('verbose', '[post /createClass] Added classroom to database')
 
-					db.get('SELECT id, name, key, permissions, tags FROM classroom WHERE name = ? AND owner = ?', [className, req.session.userId], async (err, classroom) => {
+					database.get('SELECT id, name, key, permissions, tags FROM classroom WHERE name = ? AND owner = ?', [className, req.session.userId], async (err, classroom) => {
 						try {
 							if (err) throw err
 
@@ -1074,7 +674,7 @@ app.post('/createClass', isLoggedIn, permCheck, (req, res) => {
 				}
 			})
 		} else {
-			db.get("SELECT classroom.id, classroom.name, classroom.key, classroom.permissions, classroom.tags, (CASE WHEN class_polls.pollId IS NULL THEN json_array() ELSE json_group_array(DISTINCT class_polls.pollId) END) as sharedPolls, (SELECT json_group_array(json_object('id', poll_history.id, 'class', poll_history.class, 'data', poll_history.data, 'date', poll_history.date)) FROM poll_history WHERE poll_history.class = classroom.id ORDER BY poll_history.date) as pollHistory FROM classroom LEFT JOIN class_polls ON class_polls.classId = classroom.id WHERE classroom.id = ?", [classId], async (err, classroom) => {
+			database.get("SELECT classroom.id, classroom.name, classroom.key, classroom.permissions, classroom.tags, (CASE WHEN class_polls.pollId IS NULL THEN json_array() ELSE json_group_array(DISTINCT class_polls.pollId) END) as sharedPolls, (SELECT json_group_array(json_object('id', poll_history.id, 'class', poll_history.class, 'data', poll_history.data, 'date', poll_history.date)) FROM poll_history WHERE poll_history.class = classroom.id ORDER BY poll_history.date) as pollHistory FROM classroom LEFT JOIN class_polls ON class_polls.classId = classroom.id WHERE classroom.id = ?", [classId], async (err, classroom) => {
 				try {
 					if (err) throw err
 
@@ -1134,22 +734,6 @@ app.post('/createClass', isLoggedIn, permCheck, (req, res) => {
 	}
 })
 
-// D
-
-// E
-
-// F
-
-// G
-
-// H
-
-// I
-
-// J
-
-// K
-
 // L
 // This renders the login page
 // It displays the title and the color of the login page of the formbar js
@@ -1194,7 +778,7 @@ app.post('/login', async (req, res) => {
 			logger.log('verbose', '[post /login] User is logging in')
 
 			// Get the users login in data to verify password
-			db.get('SELECT users.*, CASE WHEN shared_polls.pollId IS NULL THEN json_array() ELSE json_group_array(DISTINCT shared_polls.pollId) END as sharedPolls, CASE WHEN custom_polls.id IS NULL THEN json_array() ELSE json_group_array(DISTINCT custom_polls.id) END as ownedPolls FROM users LEFT JOIN shared_polls ON shared_polls.userId = users.id LEFT JOIN custom_polls ON custom_polls.owner = users.id WHERE users.username=?', [user.username], async (err, userData) => {
+			database.get('SELECT users.*, CASE WHEN shared_polls.pollId IS NULL THEN json_array() ELSE json_group_array(DISTINCT shared_polls.pollId) END as sharedPolls, CASE WHEN custom_polls.id IS NULL THEN json_array() ELSE json_group_array(DISTINCT custom_polls.id) END as ownedPolls FROM users LEFT JOIN shared_polls ON shared_polls.userId = users.id LEFT JOIN custom_polls ON custom_polls.owner = users.id WHERE users.username=?', [user.username], async (err, userData) => {
 				try {
 					// Check if a user with that name was not found in the database
 					if (!userData.username) {
@@ -1207,7 +791,7 @@ app.post('/login', async (req, res) => {
 					}
 
 					if (!userData.displayName) {
-						db.run("UPDATE users SET displayName = ? WHERE username = ?", [userData.username, userData.username]), (err) => {
+						database.run("UPDATE users SET displayName = ? WHERE username = ?", [userData.username, userData.username]), (err) => {
 							try {
 								if (err) throw err;
 								logger.log('verbose', '[post /login] Added displayName to database');
@@ -1289,7 +873,7 @@ app.post('/login', async (req, res) => {
 
 			let permissions = STUDENT_PERMISSIONS
 
-			db.all('SELECT API, secret, username FROM users', (err, users) => {
+			database.all('SELECT API, secret, username FROM users', (err, users) => {
 				try {
 					if (err) throw err
 
@@ -1321,7 +905,7 @@ app.post('/login', async (req, res) => {
 					} while (existingSecrets.includes(newSecret))
 
 					// Add the new user to the database
-					db.run(
+					database.run(
 						'INSERT INTO users(username, password, permissions, API, secret, displayName) VALUES(?, ?, ?, ?, ?, ?)',
 						[
 							user.username,
@@ -1337,7 +921,7 @@ app.post('/login', async (req, res) => {
 								logger.log('verbose', '[post /login] Added user to database')
 
 								// Find the user in which was just created to get the id of the user
-								db.get('SELECT * FROM users WHERE username=?', [user.username], (err, userData) => {
+								database.get('SELECT * FROM users WHERE username=?', [user.username], (err, userData) => {
 									try {
 										if (err) throw err
 
@@ -1494,7 +1078,7 @@ app.post('/oauth', (req, res) => {
 			return
 		}
 
-		db.get('SELECT * FROM users WHERE username=?', [username], (err, userData) => {
+		database.get('SELECT * FROM users WHERE username=?', [username], (err, userData) => {
 			try {
 				if (err) throw err
 
@@ -1559,7 +1143,7 @@ app.get('/previousLessons', isAuthenticated, permCheck, (req, res) => {
 	try {
 		logger.log('info', `[get /previousLessons] ip=(${req.ip}) session=(${JSON.stringify(req.session)})`)
 
-		db.all('SELECT * FROM lessons WHERE class=?', classInformation[req.session.class].className, async (err, lessons) => {
+		database.all('SELECT * FROM lessons WHERE class=?', classInformation[req.session.class].className, async (err, lessons) => {
 			try {
 				if (err) throw err
 
@@ -1633,7 +1217,7 @@ app.get('/selectClass', isLoggedIn, permCheck, (req, res) => {
 	try {
 		logger.log('info', `[get /selectClass] ip=(${req.ip}) session=(${JSON.stringify(req.session)})`)
 
-		db.all(
+		database.all(
 			'SELECT classroom.name, classroom.key FROM users JOIN classusers ON users.id = classusers.studentId JOIN classroom ON classusers.classId = classroom.id WHERE users.username=?',
 			[req.session.username],
 			(err, joinedClasses) => {
@@ -1835,21 +1419,6 @@ app.post('/student', isAuthenticated, permCheck, (req, res) => {
 	}
 })
 
-
-// T
-
-// U
-
-// V
-
-// W
-
-// X
-
-// Y
-
-// Z
-
 // 404
 app.use((req, res, next) => {
 	try {
@@ -1882,2627 +1451,6 @@ app.use((req, res, next) => {
 		})
 	}
 })
-
-let rateLimits = {}
-let userSockets = {}
-
-//Handles the websocket communications
-io.on('connection', async (socket) => {
-	try {
-		const { api } = socket.request.headers
-
-		if (api) {
-			await new Promise((resolve, reject) => {
-				db.get(
-					'SELECT id, username FROM users WHERE API=?',
-					[api],
-					(err, userData) => {
-						try {
-							if (err) throw err
-							if (!userData) {
-								logger.log('verbose', '[socket authentication] not a valid API Key')
-								throw 'Not a valid API key'
-							}
-
-							socket.request.session.api = api
-							socket.request.session.userId = userData.id
-							socket.request.session.username = userData.username
-							socket.request.session.class = getUserClass(userData.username) || 'noClass'
-
-							socket.join(`api-${socket.request.session.api}`)
-							socket.join(`class-${socket.request.session.class}`)
-
-							socket.emit('setClass', socket.request.session.class)
-
-							resolve()
-						} catch (err) {
-							reject(err)
-						}
-					}
-				)
-			}).catch((err) => {
-				if (err instanceof Error) throw err
-			})
-		} else if (socket.request.session.username) {
-			socket.join(`class-${socket.request.session.class}`)
-			socket.join(`user-${socket.request.session.username}`)
-
-			userSockets[socket.request.session.username] = socket
-		}
-	} catch (err) {
-		logger.log('error', err.stack);
-	}
-
-	function cpUpdate(classCode = socket.request.session.class) {
-		try {
-			logger.log('info', `[cpUpdate] classCode=(${classCode})`)
-
-			let classData = classInformation[classCode]
-			let cpPermissions = Math.min(
-				classData.permissions.controlPolls,
-				classData.permissions.manageStudents,
-				classData.permissions.manageClass
-			)
-
-			advancedEmitToClass('cpUpdate', classCode, { classPermissions: cpPermissions }, classData)
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	}
-
-	function vbUpdate(classCode = socket.request.session.class) {
-		try {
-			logger.log('info', `[vbUpdate] classCode=(${classCode})`)
-
-			if (!classCode) return
-			if (classCode == 'noClass') return
-
-			let classData = structuredClone(classInformation[classCode])
-			let responses = {}
-
-			// for (let [username, student] of Object.entries(classData.students)) {
-			// 	if (
-			// 		student.break == true ||
-			// 		student.classPermissions <= STUDENT_PERMISSIONS ||
-			// 		student.classPermissions >= TEACHER_PERMISSIONS
-			// 	) delete classData.students[username]
-			// }
-
-			if (Object.keys(classData.poll.responses).length > 0) {
-				for (let [resKey, resValue] of Object.entries(classData.poll.responses)) {
-					responses[resKey] = {
-						...resValue,
-						responses: 0
-					}
-				}
-
-				for (let studentData of Object.values(classData.students)) {
-					if (Array.isArray(studentData.pollRes.buttonRes)) {
-						for (let response of studentData.pollRes.buttonRes) {
-							if (
-								studentData &&
-								Object.keys(responses).includes(response)
-							) {
-								responses[response].responses++
-							}
-						}
-
-					} else if (
-						studentData &&
-						Object.keys(responses).includes(studentData.pollRes.buttonRes)
-					) {
-						responses[studentData.pollRes.buttonRes].responses++
-					}
-				}
-			}
-
-			logger.log('verbose', `[vbUpdate] status=(${classData.poll.status}) totalResponses=(${Object.keys(classData.students).length}) polls=(${JSON.stringify(responses)}) textRes=(${classData.poll.textRes}) prompt=(${classData.poll.prompt}) weight=(${classData.poll.weight}) blind=(${classData.poll.blind})`)
-
-
-			let totalResponses = 0;
-			let totalResponders = 0
-			let totalStudentsIncluded = []
-			let totalStudentsExcluded = []
-			let totalLastResponses = classData.poll.lastResponse
-
-			//Add to the included array, then add to the excluded array, then remove from the included array. Do not add the same student to either array
-			if (totalLastResponses.length > 0) {
-				totalResponses = totalLastResponses.length
-				totalStudentsIncluded = totalLastResponses
-			}
-			else {
-				for (let student of Object.values(classData.students)) {
-					if (student.classPermissions >= TEACHER_PERMISSIONS || student.classPermissions == GUEST_PERMISSIONS) continue;
-					let included = false;
-					let excluded = false;
-					// Check if the student passes the tags test
-					if (classData.poll.requiredTags.length > 0) {
-						let studentTags = student.tags.split(",");
-						if (classData.poll.requiredTags[0][0] == "0") {
-							if (classData.poll.requiredTags.slice(1).join() == student.tags) {
-								included = true;
-							} else {
-								excluded = true;
-							}
-						} else if (classData.poll.requiredTags[0][0] == "1") {
-							let correctTags = classData.poll.requiredTags.slice(1).filter(tag => studentTags.includes(tag)).length;
-							if (correctTags == classData.poll.requiredTags.length - 1) {
-								included = true;
-							} else {
-								excluded = true;
-							}
-						}
-					}
-
-					// Check if the student's checkbox was checked
-					if (classData.poll.studentBoxes.includes(student.username)) {
-						included = true;
-					} else if (classData.poll.studentBoxes.length > 0) {
-						excluded = true;
-					}
-
-					// Check if they should be in the excluded array
-					if (student.break) {
-						excluded = true;
-					}
-
-					if (classData.poll.studentIndeterminate.includes(student.username)) {
-						excluded = true;
-					}
-
-					// Update the included and excluded lists
-					if (excluded) totalStudentsExcluded.push(student.username);
-					if (included) totalStudentsIncluded.push(student.username);
-				}
-				totalStudentsIncluded = new Set(totalStudentsIncluded)
-				totalStudentsIncluded = Array.from(totalStudentsIncluded)
-				totalStudentsExcluded = new Set(totalStudentsExcluded)
-				totalStudentsExcluded = Array.from(totalStudentsExcluded)
-			}
-
-
-			totalResponses = totalStudentsIncluded.length
-			if (totalResponses == 0 && totalStudentsExcluded.length > 0) {
-				//Make total students be equal to the total number of students in the class minus the number of students who failed the perm check
-				totalResponders = Object.keys(classData.students).length - totalStudentsExcluded.length
-			}
-			else if (totalResponses == 0) {
-				totalStudentsIncluded = Object.keys(classData.students)
-				for (let i = totalStudentsIncluded.length - 1; i >= 0; i--) {
-					let student = totalStudentsIncluded[i];
-					if (classData.students[student].classPermissions >= TEACHER_PERMISSIONS || classData.students[student].classPermissions == GUEST_PERMISSIONS) {
-						totalStudentsIncluded.splice(i, 1);
-					}
-				}
-				totalResponders = totalStudentsIncluded.length
-			}
-			if (classInformation[classCode].poll.multiRes) {
-				for (let student of Object.values(classData.students)) {
-					if (student.pollRes.buttonRes.length > 1) {
-						totalResponses += student.pollRes.buttonRes.length - 1
-					}
-				}
-			} else {
-				for (let value of Object.values(classData.students)) {
-					if (value.pollRes.buttonRes != "" || value.pollRes.textRes != "") {
-						totalResponses++;
-					}
-				}
-			}
-
-			//Get rid of students whos permissions are teacher or above or guest
-			classInformation[classCode].poll.allowedResponses = totalStudentsIncluded
-			classInformation[classCode].poll.unallowedResponses = totalStudentsExcluded
-
-			advancedEmitToClass('vbUpdate', classCode, { classPermissions: CLASS_SOCKET_PERMISSIONS.vbUpdate }, {
-				status: classData.poll.status,
-				totalResponders: totalResponders,
-				totalResponses: totalResponses,
-				polls: responses,
-				textRes: classData.poll.textRes,
-				prompt: classData.poll.prompt,
-				weight: classData.poll.weight,
-				blind: classData.poll.blind,
-				time: classData.timer.time,
-				sound: classData.timer.sound,
-				active: classData.timer.active,
-				timePassed: classData.timer.timePassed,
-			})
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	}
-
-	function pollUpdate(classCode = socket.request.session.class) {
-		try {
-			logger.log('info', `[pollUpdate] classCode=(${classCode})`)
-			logger.log('verbose', `[pollUpdate] poll=(${JSON.stringify(classInformation[classCode].poll)})`)
-
-			advancedEmitToClass(
-				'pollUpdate',
-				classCode,
-				{ classPermissions: CLASS_SOCKET_PERMISSIONS.pollUpdate },
-				classInformation[socket.request.session.class].poll
-			)
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	}
-
-	function modeUpdate(classCode = socket.request.session.class) {
-		try {
-			logger.log('info', `[modeUpdate] classCode=(${classCode})`)
-			logger.log('verbose', `[modeUpdate] mode=(${classInformation[classCode].mode})`)
-
-			advancedEmitToClass(
-				'modeUpdate',
-				classCode,
-				{ classPermissions: CLASS_SOCKET_PERMISSIONS.modeUpdate },
-				classInformation[socket.request.session.class].mode
-			)
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	}
-
-	function quizUpdate(classCode = socket.request.session.class) {
-		try {
-			logger.log('info', `[quizUpdate] classCode=(${classCode})`)
-			logger.log('verbose', `[quizUpdate] quiz=(${JSON.stringify(classInformation[classCode].quiz)})`)
-
-			advancedEmitToClass(
-				'quizUpdate',
-				classCode,
-				{ classPermissions: CLASS_SOCKET_PERMISSIONS.quizUpdate },
-				classInformation[socket.request.session.class].quiz
-			)
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	}
-
-	function lessonUpdate(classCode = socket.request.session.class) {
-		try {
-			logger.log('info', `[lessonUpdate] classCode=(${classCode})`)
-			logger.log('verbose', `[lessonUpdate] lesson=(${JSON.stringify(classInformation[classCode].lesson)})`)
-
-			advancedEmitToClass(
-				'lessonUpdate',
-				classCode,
-				{ classPermissions: CLASS_SOCKET_PERMISSIONS.lessonUpdate },
-				classInformation[socket.request.session.class].lesson
-			)
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	}
-
-	function pluginUpdate(classCode = socket.request.session.class) {
-		try {
-			logger.log('info', `[pluginUpdate] classCode=(${classCode})`)
-
-			db.all(
-				'SELECT plugins.id, plugins.name, plugins.url FROM plugins JOIN classroom ON classroom.key=?',
-				[classCode],
-				(err, plugins) => {
-					try {
-						if (err) throw err
-
-						logger.log('verbose', `[pluginUpdate] plugins=(${JSON.stringify(plugins)})`)
-
-						advancedEmitToClass(
-							'pluginUpdate',
-							classCode,
-							{ classPermissions: CLASS_SOCKET_PERMISSIONS.pluginUpdate },
-							plugins
-						)
-					} catch (err) {
-						logger.log('error', err.stack);
-					}
-				}
-			)
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	}
-
-	function customPollUpdate(username) {
-		try {
-			logger.log('info', `[customPollUpdate] username=(${username})`)
-			let userSession = userSockets[username].request.session
-			let userSharedPolls = classInformation[userSession.class].students[userSession.username].sharedPolls
-			let userOwnedPolls = classInformation[userSession.class].students[userSession.username].ownedPolls
-			let userCustomPolls = Array.from(new Set(userSharedPolls.concat(userOwnedPolls)))
-			let classroomPolls = structuredClone(classInformation[userSession.class].sharedPolls)
-			let publicPolls = []
-			let customPollIds = userCustomPolls.concat(classroomPolls)
-
-			logger.log('verbose', `[customPollUpdate] userSharedPolls=(${userSharedPolls}) userOwnedPolls=(${userOwnedPolls}) userCustomPolls=(${userCustomPolls}) classroomPolls=(${classroomPolls}) publicPolls=(${publicPolls}) customPollIds=(${customPollIds})`)
-
-			db.all(
-				`SELECT * FROM custom_polls WHERE id IN(${customPollIds.map(() => '?').join(', ')}) OR public = 1 OR owner=?`,
-				[
-					...customPollIds,
-					userSession.userId
-				],
-				(err, customPollsData) => {
-					try {
-						if (err) throw err
-
-						for (let customPoll of customPollsData) {
-							customPoll.answers = JSON.parse(customPoll.answers)
-						}
-
-						customPollsData = customPollsData.reduce((newObject, customPoll) => {
-							try {
-								newObject[customPoll.id] = customPoll
-								return newObject
-							} catch (err) {
-								logger.log('error', err.stack);
-							}
-						}, {})
-
-						for (let customPoll of Object.values(customPollsData)) {
-							if (customPoll.public) {
-								publicPolls.push(customPoll.id)
-							}
-						}
-
-						logger.log('verbose', `[customPollUpdate] publicPolls=(${publicPolls}) classroomPolls=(${classroomPolls}) userCustomPolls=(${userCustomPolls}) customPollsData=(${JSON.stringify(customPollsData)})`)
-
-						io.to(`user-${username}`).emit(
-							'customPollUpdate',
-							publicPolls,
-							classroomPolls,
-							userCustomPolls,
-							customPollsData
-						)
-					} catch (err) {
-						logger.log('error', err.stack);
-					}
-				}
-			)
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	}
-
-	function classBannedUsersUpdate(classCode = socket.request.session.class) {
-		try {
-			logger.log('info', `[classBannedUsersUpdate] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[classBannedUsersUpdate] classCode=(${classCode})`)
-
-			if (!classCode || classCode == 'noClass') return
-
-			db.all('SELECT users.username FROM classroom JOIN classusers ON classusers.classId = classroom.id AND classusers.permissions = 0 JOIN users ON users.id = classusers.studentId WHERE classusers.classId=?', classInformation[socket.request.session.class].id, (err, bannedStudents) => {
-				try {
-					if (err) throw err
-
-					bannedStudents = bannedStudents.map((bannedStudent) => bannedStudent.username)
-
-					advancedEmitToClass(
-						'classBannedUsersUpdate',
-						classCode,
-						{ classPermissions: classInformation[classCode].permissions.manageStudents },
-						bannedStudents
-					)
-				} catch (err) {
-					logger.log('error', err.stack)
-				}
-			})
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	}
-
-	function classKickUser(username, classCode = socket.request.session.class) {
-		try {
-			logger.log('info', `[classKickUser] username=(${username}) classCode=(${classCode})`)
-
-			userSockets[username].leave(`class-${classCode}`)
-			classInformation.noClass.students[username] = classInformation[classCode].students[username]
-			classInformation.noClass.students[username].classPermissions = null
-			userSockets[username].request.session.class = 'noClass'
-			userSockets[username].request.session.save()
-			delete classInformation[classCode].students[username]
-
-			setClassOfApiSockets(classInformation.noClass.students[username].API, 'noClass')
-
-			logger.log('verbose', `[classKickUser] cD=(${JSON.stringify(classInformation)})`)
-
-			userSockets[username].emit('reload')
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	}
-
-	function classKickStudents(classCode) {
-		try {
-			logger.log('info', `[classKickStudents] classCode=(${classCode})`)
-
-			for (let username of Object.keys(classInformation[classCode].students)) {
-				if (classInformation[classCode].students[username].classPermissions < TEACHER_PERMISSIONS) {
-					classKickUser(username, classCode)
-				}
-			}
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	}
-
-	function logout(socket) {
-		const username = socket.request.session.username
-		const userId = socket.request.session.userId
-		const classCode = socket.request.session.class
-		const className = classInformation[classCode].className
-
-		socket.request.session.destroy((err) => {
-			try {
-				if (err) throw err
-
-				delete userSockets[username]
-				delete classInformation[classCode].students[username]
-				socket.leave(`class-${classCode}`)
-				socket.emit('reload')
-				cpUpdate(classCode)
-				vbUpdate(classCode)
-
-				db.get(
-					'SELECT * FROM classroom WHERE owner=? AND key=?',
-					[userId, classCode],
-					(err, classroom) => {
-						if (err) logger.log('error', err.stack)
-						if (classroom) endClass(classroom.key)
-					}
-				)
-			} catch (err) {
-				logger.log('error', err.stack)
-			}
-		})
-	}
-
-	async function endPoll() {
-		try {
-			logger.log('info', `[endPoll] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-
-			let data = { prompt: '', names: [], letter: [], text: [] }
-			currentPoll += 1
-
-			let dateConfig = new Date()
-			let date = `${dateConfig.getMonth() + 1} /${dateConfig.getDate()}/${dateConfig.getFullYear()}`
-
-			data.prompt = classInformation[socket.request.session.class].poll.prompt
-
-			for (const key in classInformation[socket.request.session.class].students) {
-				data.names.push(classInformation[socket.request.session.class].students[key].username)
-				data.letter.push(classInformation[socket.request.session.class].students[key].pollRes.buttonRes)
-				data.text.push(classInformation[socket.request.session.class].students[key].pollRes.textRes)
-			}
-
-			await new Promise((resolve, reject) => {
-				db.run(
-					'INSERT INTO poll_history(class, data, date) VALUES(?, ?, ?)',
-					[classInformation[socket.request.session.class].id, JSON.stringify(data), date], (err) => {
-						if (err) {
-							logger.log('error', err.stack);
-							reject(new Error(err));
-						} else {
-							logger.log('verbose', '[endPoll] saved poll to history');
-							resolve();
-						};
-					}
-				);
-			});
-
-			let latestPoll = await new Promise((resolve, reject) => {
-				db.get('SELECT * FROM poll_history WHERE class=? ORDER BY id DESC LIMIT 1', [
-					classInformation[socket.request.session.class].id
-				], (err, poll) => {
-					if (err) {
-						logger.log("error", err.stack);
-						reject(new Error(err));
-					} else resolve(poll);
-				});
-			});
-
-			latestPoll.data = JSON.parse(latestPoll.data);
-			classInformation[socket.request.session.class].pollHistory.push(latestPoll);
-
-			classInformation[socket.request.session.class].poll.status = false
-
-			logger.log('verbose', `[endPoll] classData=(${JSON.stringify(classInformation[socket.request.session.class])})`)
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	}
-
-	async function clearPoll(classCode = socket.request.session.class) {
-		if (classInformation[classCode].poll.status) await endPoll()
-
-		classInformation[classCode].poll.responses = {};
-		classInformation[classCode].poll.prompt = "";
-		classInformation[classCode].poll = {
-			status: false,
-			responses: {},
-			textRes: false,
-			prompt: "",
-			weight: 1,
-			blind: false,
-			requiredTags: [],
-			studentBoxes: [],
-			studentIndeterminate: [],
-			lastResponse: [],
-			allowedResponses: [],
-		};
-	}
-
-	async function endClass(classCode) {
-		try {
-			logger.log('info', `[endClass] classCode=(${classCode})`)
-
-			await advancedEmitToClass('endClassSound', classCode, { api: true })
-
-			for (let username of Object.keys(classInformation[classCode].students)) {
-				classKickUser(username, classCode)
-			}
-			delete classInformation[classCode]
-
-			logger.log('verbose', `[endClass] cD=(${JSON.stringify(classInformation)})`)
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	}
-
-	function getOwnedClasses(username) {
-		try {
-			logger.log('info', `[getOwnedClasses] username=(${username})`)
-
-			db.all('SELECT name, id FROM classroom WHERE owner=?',
-				[userSockets[username].request.session.userId], (err, ownedClasses) => {
-					try {
-						if (err) throw err
-
-						logger.log('info', `[getOwnedClasses] ownedClasses=(${JSON.stringify(ownedClasses)})`)
-
-						io.to(`user-${username}`).emit('getOwnedClasses', ownedClasses)
-					} catch (err) {
-						logger.log('error', err.stack);
-					}
-				}
-			)
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	}
-
-	function getPollShareIds(pollId) {
-		try {
-			logger.log('info', `[getPollShareIds] pollId=(${pollId})`)
-
-			db.all(
-				'SELECT pollId, userId, username FROM shared_polls LEFT JOIN users ON users.id = shared_polls.userId WHERE pollId=?',
-				pollId,
-				(err, userPollShares) => {
-					try {
-						if (err) throw err
-
-						db.all(
-							'SELECT pollId, classId, name FROM class_polls LEFT JOIN classroom ON classroom.id = class_polls.classId WHERE pollId=?',
-							pollId,
-							(err, classPollShares) => {
-								try {
-									if (err) throw err
-
-									logger.log('info', `[getPollShareIds] userPollShares=(${JSON.stringify(userPollShares)}) classPollShares=(${JSON.stringify(classPollShares)})`)
-
-									socket.emit('getPollShareIds', userPollShares, classPollShares)
-								} catch (err) {
-									logger.log('error', err.stack);
-								}
-							}
-						)
-					} catch (err) {
-
-					}
-				}
-			)
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	}
-
-	async function deleteCustomPolls(userId) {
-		try {
-			const customPolls = await getAll('SELECT * FROM custom_polls WHERE owner=?', userId)
-
-			if (customPolls.length == 0) return
-
-			await runQuery('DELETE FROM custom_polls WHERE userId=?', customPolls[0].userId)
-
-			for (let customPoll of customPolls) {
-				await runQuery('DELETE FROM shared_polls WHERE pollId=?', customPoll.pollId)
-			}
-		} catch (err) {
-			throw err
-		}
-	}
-
-	async function deleteClassrooms(userId) {
-		try {
-			const classrooms = await getAll('SELECT * FROM classroom WHERE owner=?', userId)
-
-			if (classrooms.length == 0) return
-
-			await runQuery('DELETE FROM classroom WHERE owner=?', classrooms[0].owner)
-
-			for (let classroom of classrooms) {
-				if (classInformation[classroom.key]) endClass(classroom.key)
-
-				await Promise.all([
-					runQuery('DELETE FROM classusers WHERE classId=?', classroom.id),
-					runQuery('DELETE FROM class_polls WHERE classId=?', classroom.id),
-					runQuery('DELETE FROM plugins WHERE classId=?', classroom.id),
-					runQuery('DELETE FROM lessons WHERE class=?', classroom.id)
-				])
-			}
-		} catch (err) {
-			throw err
-		}
-	}
-
-	function ipUpdate(type, username) {
-		try {
-			logger.log('info', `[ipUpdate] username=(${username})`)
-
-			let ipList = {}
-			if (type == 'whitelist') ipList = whitelistedIps
-			else if (type == 'blacklist') ipList = blacklistedIps
-
-			if (type) {
-				if (username) io.to(`user-${username}`).emit('ipUpdate', type, settings[`${type}Active`], ipList)
-				else io.emit('ipUpdate', type, settings[`${type}Active`], ipList)
-			} else {
-				ipUpdate('whitelist', username)
-				ipUpdate('blacklist', username)
-			}
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	}
-
-	async function reloadPageByIp(include, ip) {
-		for (let userSocket of await io.fetchSockets()) {
-			let userIp = userSocket.handshake.address
-
-			if (userIp.startsWith('::ffff:')) userIp = userIp.slice(7)
-
-			if (
-				(include &&
-					userIp.startsWith(ip)
-				) ||
-				(
-					!include &&
-					!userIp.startsWith(ip)
-				)
-			) {
-				userSocket.emit('reload')
-			}
-		}
-	}
-
-	function timer(sound, active) {
-		try {
-			let classData = classInformation[socket.request.session.class];
-
-			if (classData.timer.timeLeft <= 0) {
-				clearInterval(runningTimers[socket.request.session.class]);
-				runningTimers[socket.request.session.class] = null;
-			}
-
-			if (classData.timer.timeLeft > 0 && active) classData.timer.timeLeft--;
-
-			if (classData.timer.timeLeft <= 0 && active && sound) {
-				advancedEmitToClass('timerSound', socket.request.session.class, {
-					classPermissions: Math.max(CLASS_SOCKET_PERMISSIONS.vbTimer, classInformation[socket.request.session.class].permissions.sounds),
-					api: true
-				});
-			}
-
-			advancedEmitToClass('vbTimer', socket.request.session.class, {
-				classPermissions: CLASS_SOCKET_PERMISSIONS.vbTimer
-			}, classData.timer);
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	}
-
-	// Authentication for users and plugins to connect to formbar websockets
-	// The user must be logged in order to connect to websockets
-	socket.use(([event, ...args], next) => {
-		try {
-			let { api } = socket.request.headers
-
-			logger.log('info', `[socket authentication] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)}) api=(${api}) event=(${event})`)
-
-			if (socket.request.session.username) {
-				next()
-			} else if (api) {
-				db.get(
-					'SELECT id, username FROM users WHERE API = ?',
-					[api],
-					(err, userData) => {
-						try {
-							if (err) throw err
-							if (!userData) {
-								logger.log('verbose', '[socket authentication] not a valid API Key')
-								next(new Error('Not a valid API key'))
-								return
-							}
-
-							socket.request.session.api = api
-							socket.request.session.userId = userData.id
-							socket.request.session.username = userData.username
-							socket.request.session.class = 'noClass'
-
-							next()
-						} catch (err) {
-							logger.log('error', err.stack)
-						}
-					}
-				)
-			} else if (event == 'reload') {
-				next()
-			} else {
-				logger.log('info', '[socket authentication] Missing username or api')
-				next(new Error('Missing API key'))
-			}
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	//rate limiter
-	socket.use(([event, ...args], next) => {
-		try {
-			const username = socket.request.session.username
-			const currentTime = Date.now()
-			const limit = 5
-			const timeFrame = 5000
-			const blockTime = 5000
-			const limitedRequests = ['pollResp', 'help', 'break']
-
-			logger.log('info', `[rate limiter] username=(${username}) currentTime=(${currentTime})`)
-
-			if (!rateLimits[username]) {
-				rateLimits[username] = {}
-			}
-
-			const userRequests = rateLimits[username]
-
-			if (!limitedRequests.includes(event)) {
-				next()
-				return
-			}
-
-			userRequests[event] = userRequests[event] || []
-
-			userRequests[event] = userRequests[event].filter((timestamp) => currentTime - timestamp < timeFrame)
-
-			logger.log('verbose', `[rate limiter] userRequests=(${JSON.stringify(userRequests)})`)
-
-			if (userRequests[event].length >= limit) {
-				socket.emit('message', `You are being rate limited. Please try again in a ${blockTime / 1000} seconds.`)
-				next(new Error('Rate limited'))
-				setTimeout(() => {
-					try {
-						userRequests[event].shift()
-					} catch (err) {
-						logger.log('error', err.stack);
-					}
-				}, blockTime)
-			} else {
-				userRequests[event].push(currentTime)
-				next()
-			}
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	})
-
-	// permission check
-	socket.use(async ([event, ...args], next) => {
-		try {
-			let username = socket.request.session.username
-			let classCode = socket.request.session.class
-
-			logger.log('info', `[socket permission check] Event=(${event}), Username=(${username}), ClassCod=(${classCode})`)
-
-			if (!classInformation[classCode]) {
-				logger.log('info', '[socket permission check] Class does not exist')
-				socket.emit('message', 'Class does not exist')
-				return
-			}
-			if (!classInformation[classCode].students[username]) {
-				logger.log('info', '[socket permission check] User is not logged in')
-				socket.emit('message', 'User is not logged in')
-				return
-			}
-
-			if (
-				GLOBAL_SOCKET_PERMISSIONS[event] &&
-				classInformation[classCode].students[username].permissions >= GLOBAL_SOCKET_PERMISSIONS[event]
-			) {
-				logger.log('info', '[socket permission check] Global socket permission check passed')
-				next()
-			} else if (
-				CLASS_SOCKET_PERMISSIONS[event] &&
-				classInformation[classCode].students[username].classPermissions >= CLASS_SOCKET_PERMISSIONS[event]
-			) {
-				logger.log('info', '[socket permission check] Class socket permission check passed')
-				next()
-			} else if (
-				CLASS_SOCKET_PERMISSION_SETTINGS[event] &&
-				classInformation[classCode].permissions[CLASS_SOCKET_PERMISSION_SETTINGS[event]] &&
-				classInformation[classCode].students[username].classPermissions >= classInformation[classCode].permissions[CLASS_SOCKET_PERMISSION_SETTINGS[event]]
-			) {
-				logger.log('info', '[socket permission check] Class socket permission settings check passed')
-				next()
-			} else {
-				if (!PASSIVE_SOCKETS.includes(event)) {
-					logger.log('info', `[socket permission check] User does not have permission to use ${camelCaseToNormal(event)}`)
-					socket.emit('message', `You do not have permission to use ${camelCaseToNormal(event)}.`)
-				}
-			}
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	// /poll websockets for updating the database
-	socket.on('pollResp', (res, textRes, resWeight, resLength) => {
-		try {
-			logger.log('info', `[pollResp] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[pollResp] res=(${res}) textRes=(${textRes}) resWeight=(${resWeight}) resLength=(${resLength})`)
-
-			if (
-				classInformation[socket.request.session.class].students[socket.request.session.username].pollRes.buttonRes != res ||
-				classInformation[socket.request.session.class].students[socket.request.session.username].pollRes.textRes != textRes
-			) {
-				if (res == 'remove')
-					advancedEmitToClass('removePollSound', socket.request.session.class, { api: true })
-				else
-					advancedEmitToClass('pollSound', socket.request.session.class, { api: true })
-			}
-
-			classInformation[socket.request.session.class].students[socket.request.session.username].pollRes.buttonRes = res
-			classInformation[socket.request.session.class].students[socket.request.session.username].pollRes.textRes = textRes
-			classInformation[socket.request.session.class].students[socket.request.session.username].pollRes.time = new Date()
-
-
-			for (let i = 0; i < resLength; i++) {
-				if (res) {
-					let calcWeight = classInformation[socket.request.session.class].poll.weight * resWeight
-					classInformation[socket.request.session.class].students[socket.request.session.username].pogMeter += calcWeight
-					if (classInformation[socket.request.session.class].students[socket.request.session.username].pogMeter >= 25) {
-						db.get('SELECT digipogs FROM classusers WHERE studentId=?', [classInformation[socket.request.session.class].students[socket.request.session.username].id], (err, data) => {
-							try {
-								if (err) throw err
-
-								db.run('UPDATE classusers SET digiPogs=? WHERE studentId=?', [data + 1, classInformation[socket.request.session.class].students[socket.request.session.username].id], (err) => {
-									try {
-										if (err) throw err
-
-										logger.log('verbose', `[pollResp] Added 1 digipog to ${socket.request.session.username}`)
-									} catch (err) {
-										logger.log('error', err.stack);
-									}
-								})
-							} catch (err) {
-								logger.log('error', err.stack);
-							}
-						})
-						classInformation[socket.request.session.class].students[socket.request.session.username].pogMeter = 0
-					}
-				}
-			}
-			logger.log('verbose', `[pollResp] user=(${classInformation[socket.request.session.class].students[socket.request.session.username]})`)
-
-			cpUpdate()
-			vbUpdate()
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	})
-
-	// Changes Permission of user. Takes which user and the new permission level
-	socket.on('classPermChange', (user, newPerm) => {
-		try {
-			logger.log('info', `[classPermChange] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[classPermChange] user=(${user}) newPerm=(${newPerm})`)
-			classInformation[socket.request.session.class].students[user].classPermissions = newPerm
-
-			db.run('UPDATE classusers SET permissions=? WHERE classId=? AND studentId=?', [
-				newPerm,
-				classInformation[socket.request.session.class].id,
-				classInformation[socket.request.session.class].students[user].id
-			])
-
-			logger.log('verbose', `[classPermChange] user=(${JSON.stringify(classInformation[socket.request.session.class].students[user])})`)
-			io.to(`user-${user}`).emit('reload')
-
-			//cpUpdate()
-			//Commented Out to fix Issue #231 checkbox 14, tags not updating when permissions are changed and page is not refreashed
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	})
-
-	socket.on('permChange', async (username, newPerm) => {
-		try {
-			newPerm = Number(newPerm)
-
-			logger.log('info', `[permChange] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[permChange] user=(${username}) newPerm=(${newPerm})`)
-
-			let classCode = getUserClass(username)
-			if (classCode instanceof Error) throw classCode
-
-			if (classCode) {
-				classInformation[classCode].students[username].permissions = newPerm
-
-				if (
-					classInformation[classCode].students[username].permissions < TEACHER_PERMISSIONS &&
-					Object.keys(classInformation[classCode].students)[0] == username
-				) {
-					endClass(classCode)
-				}
-
-				io.to(`user-${username}`).emit('reload')
-			}
-
-			db.run('UPDATE users SET permissions=? WHERE username=?', [newPerm, username])
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	})
-
-	// Starts a new poll. Takes the number of responses and whether or not their are text responses
-	socket.on('startPoll', async (resNumber, resTextBox, pollPrompt, polls, blind, weight, tags, boxes, indeterminate, lastResponse, multiRes) => {
-		try {
-			logger.log('info', `[startPoll] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[startPoll] resNumber=(${resNumber}) resTextBox=(${resTextBox}) pollPrompt=(${pollPrompt}) polls=(${JSON.stringify(polls)}) blind=(${blind}) weight=(${weight}) tags=(${tags})`)
-
-			await clearPoll()
-			let generatedColors = generateColors(resNumber)
-			logger.log('verbose', `[pollResp] user=(${classInformation[socket.request.session.class].students[socket.request.session.username]})`)
-			if (generatedColors instanceof Error) throw generatedColors
-
-			classInformation[socket.request.session.class].mode = 'poll'
-			classInformation[socket.request.session.class].poll.blind = blind
-			classInformation[socket.request.session.class].poll.status = true
-			if (tags) {
-				classInformation[socket.request.session.class].poll.requiredTags = tags
-			}
-			else {
-				classInformation[socket.request.session.class].poll.requiredTags = []
-			}
-			if (boxes) {
-				classInformation[socket.request.session.class].poll.studentBoxes = boxes
-			}
-			else {
-				classInformation[socket.request.session.class].poll.studentBoxes = []
-			}
-			if (indeterminate) {
-				classInformation[socket.request.session.class].poll.studentIndeterminate = indeterminate
-			}
-			else {
-				classInformation[socket.request.session.class].poll.studentIndeterminate = []
-			}
-			if (lastResponse) {
-				classInformation[socket.request.session.class].poll.lastResponse = lastResponse
-			}
-			else {
-				classInformation[socket.request.session.class].poll.lastResponse = []
-			}
-
-
-
-
-
-
-			// Creates an object for every answer possible the teacher is allowing
-			for (let i = 0; i < resNumber; i++) {
-				let letterString = 'abcdefghijklmnopqrstuvwxyz'
-				let answer = letterString[i]
-				let weight = 1
-				let color = generatedColors[i]
-
-				if (polls[i].answer)
-					answer = polls[i].answer
-				if (polls[i].weight)
-					weight = polls[i].weight
-				if (polls[i].color)
-					color = polls[i].color
-
-				classInformation[socket.request.session.class].poll.responses[answer] = {
-					answer: answer,
-					weight: weight,
-					color: color
-				}
-			}
-
-			classInformation[socket.request.session.class].poll.weight = weight
-			classInformation[socket.request.session.class].poll.textRes = resTextBox
-			classInformation[socket.request.session.class].poll.prompt = pollPrompt
-			classInformation[socket.request.session.class].poll.multiRes = multiRes
-
-			for (var key in classInformation[socket.request.session.class].students) {
-				classInformation[socket.request.session.class].students[key].pollRes.buttonRes = ''
-				classInformation[socket.request.session.class].students[key].pollRes.textRes = ''
-			}
-
-			logger.log('verbose', `[startPoll] classData=(${JSON.stringify(classInformation[socket.request.session.class])})`)
-
-			pollUpdate()
-			vbUpdate()
-			cpUpdate()
-			socket.emit('startPoll')
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	})
-
-	// End the current poll. Does not take any arguments
-	socket.on('clearPoll', async () => {
-		try {
-			await clearPoll();
-			//adds data to the previous poll answers table upon clearing the poll
-			for (var student of Object.values(classInformation[socket.request.session.class].students)) {
-				if (student.classPermissions != 5) {
-					var currentPollId = classInformation[socket.request.session.class].pollHistory[currentPoll].id
-					for (let i = 0; i < student.pollRes.buttonRes.length; i++) {
-						var studentRes = student.pollRes.buttonRes[i]
-						var studentId = student.id
-						db.run('INSERT INTO poll_answers(pollId, userId, buttonResponse) VALUES(?, ?, ?)', [currentPollId, studentId, studentRes], (err) => {
-							if (err) {
-								logger.log('error', err.stack)
-							}
-						})
-					}
-					var studentTextRes = student.pollRes.textRes
-					var studentId = student.id
-					db.run('INSERT INTO poll_answers(pollId, userId, textResponse) VALUES(?, ?, ?)', [currentPollId, studentId, studentTextRes], (err) => {
-						if (err) {
-							logger.log('error', err.stack)
-						}
-					})
-				}
-			}
-
-			pollUpdate();
-			vbUpdate();
-			cpUpdate();
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	})
-
-	socket.on('endPoll', async () => {
-		try {
-			await endPoll();
-			pollUpdate();
-			cpUpdate();
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	})
-
-	socket.on('pollUpdate', () => {
-		logger.log('info', `[pollUpdate] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-		pollUpdate()
-	})
-
-	socket.on('modeUpdate', () => {
-		logger.log('info', `[modeUpdate] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-
-		modeUpdate()
-	})
-
-	socket.on('quizUpdate', () => {
-		logger.log('info', `[quizUpdate] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-
-		quizUpdate()
-	})
-
-	socket.on('lessonUpdate', () => {
-		logger.log('info', `[lessonUpdate] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-
-		lessonUpdate()
-	})
-
-	// Sends poll and student response data to client side virtual bar
-	socket.on('vbUpdate', () => {
-		logger.log('info', `[vbUpdate] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-
-		vbUpdate()
-	})
-
-	socket.on('customPollUpdate', () => {
-		logger.log('info', `[customPollUpdate] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-
-		customPollUpdate(socket.request.session.username)
-	})
-
-	socket.on('savePoll', (poll, id) => {
-		try {
-			logger.log('info', `[savePoll] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[savePoll] poll=(${JSON.stringify(poll)}) id=(${id})`)
-
-			let userId = socket.request.session.userId
-
-			if (id) {
-				db.get('SELECT * FROM custom_polls WHERE id=?', [id], (err, poll) => {
-					try {
-						if (err) throw err
-
-						if (userId != poll.owner) {
-							socket.emit('message', 'You do not have permission to edit this poll.')
-							return
-						}
-
-						db.run('UPDATE custom_polls SET name=?, prompt=?, answers=?, textRes=?, blind=?, weight=?, public=? WHERE id=?', [
-							poll.name,
-							poll.prompt,
-							JSON.stringify(poll.answers),
-							poll.textRes,
-							poll.blind,
-							poll.weight,
-							poll.public,
-							id
-						], (err) => {
-							try {
-								if (err) throw err
-
-								socket.emit('message', 'Poll saved successfully!')
-								customPollUpdate(socket.request.session.username)
-							} catch (err) {
-								logger.log('error', err.stack);
-							}
-						})
-					} catch (err) {
-						logger.log('error', err.stack);
-					}
-				})
-			} else {
-				db.get('SELECT seq AS nextPollId from sqlite_sequence WHERE name = "custom_polls"', (err, nextPollId) => {
-					try {
-						if (err) throw err
-						if (!nextPollId) logger.log('critical', '[savePoll] nextPollId not found')
-
-						nextPollId = nextPollId.nextPollId + 1
-
-						db.run('INSERT INTO custom_polls (owner, name, prompt, answers, textRes, blind, weight, public) VALUES (?, ?, ?, ?, ?, ?, ?, ?)', [
-							userId,
-							poll.name,
-							poll.prompt,
-							JSON.stringify(poll.answers),
-							poll.textRes,
-							poll.blind,
-							poll.weight,
-							poll.public
-						], (err) => {
-							try {
-								if (err) throw err
-
-								classInformation[socket.request.session.class].students[socket.request.session.username].ownedPolls.push(nextPollId)
-								socket.emit('message', 'Poll saved successfully!')
-								customPollUpdate(socket.request.session.username)
-							} catch (err) {
-								logger.log('error', err.stack);
-							}
-						})
-					} catch (err) {
-						logger.log('error', err.stack);
-					}
-				})
-			}
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	})
-
-	socket.on('deletePoll', (pollId) => {
-		try {
-			let userId = socket.request.session.userId
-
-			logger.log('info', `[deletePoll] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)}) pollId=(${pollId})`)
-			logger.log('info', `[deletePoll] pollId=(${pollId})`)
-
-			if (!pollId) {
-				socket.emit('message', 'No poll is selected.')
-				return
-			}
-
-			db.get('SELECT * FROM custom_polls WHERE id=?', pollId, async (err, poll) => {
-				try {
-					if (err) throw err
-
-					logger.log('info', `[deletePoll] poll=(${JSON.stringify(poll)})`)
-
-					if (userId != poll.owner) {
-						logger.log('info', '[deletePoll] not owner')
-						socket.emit('message', 'You do not have permission to delete this poll.')
-						return
-					}
-
-					await runQuery('BEGIN TRANSACTION')
-
-					await Promise.all([
-						runQuery('DELETE FROM custom_polls WHERE id=?', pollId),
-						runQuery('DELETE FROM shared_polls WHERE pollId=?', pollId),
-						runQuery('DELETE FROM class_polls WHERE pollId=?', pollId),
-					]).catch(async (err) => {
-						await runQuery('ROLLBACK')
-						throw err
-					})
-
-					await runQuery('COMMIT')
-
-					for (let classroom of Object.values(classInformation)) {
-						let updatePolls = false
-
-						if (classroom.sharedPolls) {
-							if (classroom.sharedPolls.includes(pollId)) {
-								classroom.sharedPolls.splice(classroom.sharedPolls.indexOf(pollId), 1)
-								updatePolls = true
-							}
-						}
-
-						for (let user of Object.values(classroom.students)) {
-							if (user.sharedPolls.includes(pollId)) {
-								user.sharedPolls.splice(user.sharedPolls.indexOf(pollId), 1)
-								updatePolls = true
-							}
-
-							if (user.ownedPolls.includes(pollId)) {
-								user.ownedPolls.splice(user.ownedPolls.indexOf(pollId), 1)
-								updatePolls = true
-							}
-
-							if (updatePolls)
-								customPollUpdate(user.username)
-						}
-					}
-
-					logger.log('info', '[deletePoll] deleted')
-					socket.emit('message', 'Poll deleted successfully!')
-				} catch (err) {
-					logger.log('error', err.stack);
-				}
-			})
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	})
-
-	socket.on('setPublicPoll', (pollId, value) => {
-		try {
-			logger.log('info', `[setPublicPoll] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[setPublicPoll] pollId=(${pollId}) value=(${value})`)
-
-			db.run('UPDATE custom_polls set public=? WHERE id=?', [value, pollId], (err) => {
-				try {
-					if (err) throw err
-
-					for (let userSocket of Object.values(userSockets)) {
-						customPollUpdate(userSocket.request.session.username)
-					}
-				} catch (err) {
-					logger.log('error', err.stack);
-				}
-			})
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	})
-
-	socket.on('sharePollToUser', (pollId, username) => {
-		try {
-			logger.log('info', `[sharePollToUser] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[sharePollToUser] pollId=(${pollId}) username=(${username})`)
-
-			db.get('SELECT * FROM users WHERE username=?', username, (err, user) => {
-				try {
-					if (err) throw err
-
-					if (!user) {
-						logger.log('info', 'User does not exist')
-						socket.emit('message', 'User does not exist')
-						return
-					}
-
-					db.get('SELECT * FROM custom_polls WHERE id=?', pollId, (err, poll) => {
-						try {
-							if (err) throw err
-
-							if (!poll) {
-								logger.log('critical', 'Poll does not exist')
-								socket.emit('message', 'Poll does not exist (Please contact the programmer)')
-								return
-							}
-
-							let name = 'Unnamed Poll'
-							if (poll.name) name = poll.name
-							else if (poll.prompt) name = poll.prompt
-
-							db.get(
-								'SELECT * FROM shared_polls WHERE pollId=? AND userId=?',
-								[pollId, user.id],
-								(err, sharePoll) => {
-									try {
-										if (err) throw err
-
-										if (sharePoll) {
-											socket.emit('message', `${name} is Already Shared with ${username}`)
-											return
-										}
-
-										db.run(
-											'INSERT INTO shared_polls (pollId, userId) VALUES (?, ?)',
-											[pollId, user.id],
-											async (err) => {
-												try {
-													if (err) throw err
-
-													socket.emit('message', `Shared ${name} with ${username}`)
-
-													getPollShareIds(pollId)
-
-													let classCode = getUserClass(username)
-
-													if (classCode instanceof Error) throw classCode
-													if (!classCode) return
-
-													classInformation[classCode].students[user.username].sharedPolls.push(pollId)
-
-													customPollUpdate(username)
-												} catch (err) {
-													logger.log('error', err.stack);
-												}
-											}
-										)
-									} catch (err) {
-										logger.log('error', err.stack);
-									}
-								}
-							)
-						} catch (err) {
-							logger.log('error', err.stack);
-						}
-					})
-				} catch (err) {
-					logger.log('error', err.stack);
-				}
-			})
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	})
-
-	socket.on('removeUserPollShare', (pollId, userId) => {
-		try {
-			logger.log('info', `[removeUserPollShare] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[removeUserPollShare] pollId=(${pollId}) userId=(${userId})`)
-
-			db.get(
-				'SELECT * FROM shared_polls WHERE pollId=? AND userId=?',
-				[pollId, userId],
-				(err, pollShare) => {
-					try {
-						if (err) throw err
-
-						if (!pollShare) {
-							logger.log('critical', '[removeUserPollShare] Poll is not shared to this user')
-							socket.emit('message', 'Poll is not shared to this user')
-							return
-						}
-
-						db.run(
-							'DELETE FROM shared_polls WHERE pollId=? AND userId=?',
-							[pollId, userId],
-							(err) => {
-								try {
-									if (err) throw err
-
-									socket.emit('message', 'Successfully unshared user')
-									getPollShareIds(pollId)
-
-									db.get('SELECT * FROM users WHERE id=?', userId, async (err, user) => {
-										try {
-											if (err) throw err
-
-											if (!user) {
-												logger.log('critical', '[removeUserPollShare] User does not exist')
-												socket.emit('message', 'User does not exist')
-												return
-											}
-
-											let classCode = getUserClass(user.username)
-
-											if (classCode instanceof Error) throw classCode
-											if (!classCode) return
-
-											let sharedPolls = classInformation[classCode].students[user.username].sharedPolls
-											sharedPolls.splice(sharedPolls.indexOf(pollId), 1)
-											customPollUpdate(user.username)
-										} catch (err) {
-											logger.log('error', err.stack);
-										}
-									})
-								} catch (err) {
-									logger.log('error', err.stack);
-								}
-							}
-						)
-					} catch (err) {
-						logger.log('error', err.stack);
-					}
-				}
-			)
-		} catch (err) {
-			logger.log('error', err.stack);
-		}
-	})
-
-	socket.on('getPollShareIds', (pollId) => {
-		logger.log('info', `[getPollShareIds] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-		logger.log('info', `[getPollShareIds] pollId=(${pollId})`)
-
-		getPollShareIds(pollId)
-	})
-
-	socket.on('sharePollToClass', (pollId, classCode) => {
-		try {
-			logger.log('info', `[sharePollToClass] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[sharePollToClass] pollId=(${pollId}) classCode=(${classCode})`)
-
-			db.get('SELECT * FROM classroom WHERE key=?', classCode, (err, classroom) => {
-				try {
-					if (err) throw err
-
-					if (!classroom) {
-						socket.emit('message', 'There is no class with that code.')
-						return
-					}
-
-					db.get('SELECT * FROM custom_polls WHERE id=?', pollId, (err, poll) => {
-						try {
-							if (err) throw err
-
-							if (!poll) {
-								logger.log('critical', 'Poll does not exist (Please contact the programmer)')
-								socket.emit('message', 'Poll does not exist (Please contact the programmer)')
-								return
-							}
-
-							let name = 'Unnamed Poll'
-							if (poll.name) name = poll.name
-							else if (poll.prompt) name = poll.prompt
-
-							db.get(
-								'SELECT * FROM class_polls WHERE pollId=? AND classId=?',
-								[pollId, classroom.id],
-								(err, sharePoll) => {
-									try {
-										if (err) throw err
-
-										if (sharePoll) {
-											socket.emit('message', `${name} is Already Shared with that class`)
-											return
-										}
-
-										db.run(
-											'INSERT INTO class_polls (pollId, classId) VALUES (?, ?)',
-											[pollId, classroom.id],
-											async (err) => {
-												try {
-													if (err) throw err
-
-													socket.emit('message', `Shared ${name} with that class`)
-
-													getPollShareIds(pollId)
-
-													classInformation[classCode].sharedPolls.push(pollId)
-													for (let username of Object.keys(classInformation[classCode].students)) {
-														customPollUpdate(username)
-													}
-												} catch (err) {
-													logger.log('error', err.stack)
-												}
-											}
-										)
-									} catch (err) {
-										logger.log('error', err.stack)
-									}
-								}
-							)
-						} catch (err) {
-							logger.log('error', err.stack)
-						}
-					})
-				} catch (err) {
-					logger.log('error', err.stack)
-				}
-			})
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	socket.on('removeClassPollShare', (pollId, classId) => {
-		try {
-			logger.log('info', `[removeClassPollShare] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[removeClassPollShare] pollId=(${pollId}) classId=(${classId})`)
-
-			db.get(
-				'SELECT * FROM class_polls WHERE pollId=? AND classId=?',
-				[pollId, classId],
-				(err, pollShare) => {
-					try {
-						if (err) throw err
-
-						if (!pollShare) {
-							socket.emit('message', 'Poll is not shared to this class')
-							return
-						}
-
-						db.run(
-							'DELETE FROM class_polls WHERE pollId=? AND classId=?',
-							[pollId, classId],
-							(err) => {
-								try {
-									if (err) throw err
-
-									socket.emit('message', 'Successfully unshared class')
-									getPollShareIds(pollId)
-
-									db.get('SELECT * FROM classroom WHERE id=?', classId, async (err, classroom) => {
-										try {
-											if (err) throw err
-
-											if (!classroom) {
-												logger.log('critical', 'Classroom does not exist')
-												return
-											}
-											if (!classInformation[classroom.key]) return
-
-											let sharedPolls = classInformation[classroom.key].sharedPolls
-											sharedPolls.splice(sharedPolls.indexOf(pollId), 1)
-											for (let username of Object.keys(classInformation[classroom.key].students)) {
-												customPollUpdate(username)
-											}
-										} catch (err) {
-											logger.log('error', err.stack);
-										}
-									})
-								} catch (err) {
-									logger.log('error', err.stack)
-								}
-							}
-						)
-					} catch (err) {
-						logger.log('error', err.stack)
-					}
-				}
-			)
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	// Sends a help ticket
-	socket.on('help', (reason) => {
-		try {
-			logger.log('info', `[help] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-
-			let time = new Date();
-
-			logger.log('info', `[help] reason=(${reason}) time=(${time})`)
-
-			let student = classInformation[socket.request.session.class].students[socket.request.session.username]
-
-			if (student.help.reason != reason) {
-				advancedEmitToClass('helpSound', socket.request.session.class, { api: true })
-			}
-
-			student.help = { reason: reason, time: time }
-
-			logger.log('verbose', `[help] user=(${JSON.stringify(student)}`)
-
-			cpUpdate()
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	// Sends a break ticket
-	socket.on('requestBreak', (reason) => {
-		try {
-			logger.log('info', `[requestBreak] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[requestBreak] reason=(${reason})`)
-
-			let student = classInformation[socket.request.session.class].students[socket.request.session.username]
-
-			if (!student.break != reason)
-				advancedEmitToClass('breakSound', socket.request.session.class, { api: true })
-
-			student.break = reason
-
-			logger.log('verbose', `[requestBreak] user=(${JSON.stringify(classInformation[socket.request.session.class].students[socket.request.session.username])})`)
-
-			cpUpdate()
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	// Approves the break ticket request
-	socket.on('approveBreak', (breakApproval, username) => {
-		try {
-			logger.log('info', `[approveBreak] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[approveBreak] breakApproval=(${breakApproval}) username=(${username})`)
-
-			let student = classInformation[socket.request.session.class].students[username]
-			student.break = breakApproval
-
-			logger.log('verbose', `[approveBreak] user=(${JSON.stringify(classInformation[socket.request.session.class].students[username])})`)
-
-			if (breakApproval) io.to(`user-${username}`).emit('break')
-			cpUpdate()
-			vbUpdate()
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	// Ends the break
-	socket.on('endBreak', () => {
-		try {
-			logger.log('info', `[endBreak] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-
-			let student = classInformation[socket.request.session.class].students[socket.request.session.username]
-			student.break = false
-
-			logger.log('verbose', `[endBreak] user=(${JSON.stringify(classInformation[socket.request.session.class].students[socket.request.session.username])})`)
-
-			cpUpdate()
-			vbUpdate()
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	// Kicks a user from the class
-	socket.on('classKickUser', (username) => {
-		try {
-			logger.log('info', `[classKickUser] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[classKickUser] username=(${username})`)
-
-			const classCode = socket.request.session.class
-			classKickUser(username, classCode)
-			advancedEmitToClass('leaveSound', classCode, { api: true })
-			cpUpdate(classCode)
-			vbUpdate(classCode)
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	// Deletes all students from the class
-	socket.on('classKickStudents', () => {
-		try {
-			logger.log('info', `[classKickStudents] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-
-			const classCode = socket.request.session.class
-			classKickStudents(classCode)
-			advancedEmitToClass('kickStudentsSound', classCode, { api: true })
-			cpUpdate(classCode)
-			vbUpdate(classCode)
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	socket.on('leaveClass', () => {
-		try {
-			logger.log('info', `[leaveClass] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-
-			const userId = socket.request.session.userId
-			const username = socket.request.session.username
-			const classCode = socket.request.session.class
-			classKickUser(username, classCode)
-			advancedEmitToClass('leaveSound', classCode, { api: true })
-			cpUpdate(classCode)
-			vbUpdate(classCode)
-
-			db.get(
-				'SELECT * FROM classroom WHERE owner=? AND key=?',
-				[userId, classCode],
-				(err, classroom) => {
-					if (err) logger.log('error', err.stack)
-					else if (classroom) endClass(classroom.key)
-				}
-			)
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	socket.on('logout', () => {
-		try {
-			logger.log('info', `[logout] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-
-			logout(socket)
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	socket.on('endClass', () => {
-		try {
-			logger.log('info', `[endClass] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-
-			const userId = socket.request.session.userId
-			const classCode = socket.request.session.class
-
-			db.get(
-				'SELECT * FROM classroom WHERE owner=? AND key=?',
-				[userId, classCode],
-				(err, classroom) => {
-					if (err) logger.log('error', err.stack)
-					else if (classroom) endClass(classroom.key)
-				}
-			)
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	socket.on('deleteClass', (classId) => {
-		try {
-			logger.log('info', `[deleteClass] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[deleteClass] classId=(${classId})`)
-
-			db.get('SELECT * FROM classroom WHERE id=?', classId, (err, classroom) => {
-				try {
-					if (err) throw err
-
-					if (classroom) {
-						if (classInformation[classroom.key]) endClass(classroom.key)
-
-						db.run('DELETE FROM classroom WHERE id=?', classroom.id)
-						db.run('DELETE FROM classusers WHERE classId=?', classroom.id)
-						db.run('DELETE FROM poll_history WHERE class=?', classroom.id)
-					}
-
-					getOwnedClasses(socket.request.session.username)
-				} catch (err) {
-					logger.log('error', err.stack)
-				}
-			})
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	socket.on('managerUpdate', () => {
-		managerUpdate()
-	})
-
-	// Updates and stores poll history
-	socket.on('cpUpdate', () => {
-		logger.log('info', `[cpUpdate] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-
-		cpUpdate();
-	})
-
-	// Displays previous polls
-	socket.on('previousPollDisplay', (pollIndex) => {
-		try {
-			logger.log('info', `[previousPollDisplay] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[previousPollDisplay] pollIndex=(${pollIndex})`)
-
-			advancedEmitToClass(
-				'previousPollData',
-				socket.request.session.class,
-				{ classPermissions: classInformation[socket.request.session.class].permissions.controlPolls },
-				classInformation[socket.request.session.class].pollHistory[pollIndex].data
-			)
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	// Moves to the next step
-	socket.on('doStep', (index) => {
-		try {
-			logger.log('info', `[doStep] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[doStep] index=(${index})`)
-
-			// send reload to whole class
-			socket.broadcast.to(socket.request.session.class).emit('reload')
-			classInformation[socket.request.session.class].currentStep++
-
-			if (classInformation[socket.request.session.class].steps[index] !== undefined) {
-				// Creates a poll based on the step data
-				if (classInformation[socket.request.session.class].steps[index].type == 'poll') {
-					classInformation[socket.request.session.class].mode = 'poll'
-
-					if (classInformation[socket.request.session.class].poll.status == true) {
-						classInformation[socket.request.session.class].poll.responses = {}
-						classInformation[socket.request.session.class].poll.prompt = ''
-						classInformation[socket.request.session.class].poll.status = false
-					};
-
-					classInformation[socket.request.session.class].poll.status = true
-					// Creates an object for every answer possible the teacher is allowing
-					for (let i = 0; i < classInformation[socket.request.session.class].steps[index].responses; i++) {
-						if (classInformation[socket.request.session.class].steps[index].labels[i] == '' || classInformation[socket.request.session.class].steps[index].labels[i] == null) {
-							let letterString = 'abcdefghijklmnopqrstuvwxyz'
-							classInformation[socket.request.session.class].poll.responses[letterString[i]] = { answer: 'Answer ' + letterString[i], weight: 1 }
-						} else {
-							classInformation[socket.request.session.class].poll.responses[classInformation[socket.request.session.class].steps[index].labels[i]] = { answer: classInformation[socket.request.session.class].steps[index].labels[i], weight: classInformation[socket.request.session.class].steps[index].weights[i] }
-						}
-					}
-					classInformation[socket.request.session.class].poll.textRes = false
-					classInformation[socket.request.session.class].poll.prompt = classInformation[socket.request.session.class].steps[index].prompt
-					// Creates a new quiz based on step data
-				} else if (classInformation[socket.request.session.class].steps[index].type == 'quiz') {
-					classInformation[socket.request.session.class].mode = 'quiz'
-					questions = classInformation[socket.request.session.class].steps[index].questions
-					let quiz = new Quiz(questions.length, 100)
-					quiz.questions = questions
-					classInformation[socket.request.session.class].quiz = quiz
-					// Creates lesson based on step data
-				} else if (classInformation[socket.request.session.class].steps[index].type == 'lesson') {
-					classInformation[socket.request.session.class].mode = 'lesson'
-					let lesson = new Lesson(classInformation[socket.request.session.class].steps[index].date, classInformation[socket.request.session.class].steps[index].lesson)
-					classInformation[socket.request.session.class].lesson = lesson
-					db.run('INSERT INTO lessons(class, content, date) VALUES(?, ?, ?)',
-						[classInformation[socket.request.session.class].className, JSON.stringify(classInformation[socket.request.session.class].lesson), classInformation[socket.request.session.class].lesson.date], (err) => {
-							if (err) logger.log('error', err.stack)
-						}
-					)
-					classInformation[socket.request.session.class].poll.textRes = false
-					classInformation[socket.request.session.class].poll.prompt = classInformation[socket.request.session.class].steps[index].prompt
-					// Check this later, there's already a quiz if statement
-				} else if (classInformation[socket.request.session.class].steps[index].type == 'quiz') {
-					questions = classInformation[socket.request.session.class].steps[index].questions
-					quiz = new Quiz(questions.length, 100)
-					quiz.questions = questions
-					classInformation[socket.request.session.class].quiz = quiz
-					// Check this later, there's already a lesson if statement
-				} else if (classInformation[socket.request.session.class].steps[index].type == 'lesson') {
-					let lesson = new Lesson(classInformation[socket.request.session.class].steps[index].date, classInformation[socket.request.session.class].steps[index].lesson)
-					classInformation[socket.request.session.class].lesson = lesson
-					db.run('INSERT INTO lessons(class, content, date) VALUES(?, ?, ?)',
-						[classInformation[socket.request.session.class].className, JSON.stringify(classInformation[socket.request.session.class].lesson), classInformation[socket.request.session.class].lesson.date], (err) => {
-							if (err) logger.log('error', err.stack)
-						}
-					)
-				}
-
-				pollUpdate()
-				modeUpdate()
-				quizUpdate()
-				lessonUpdate()
-			} else {
-				classInformation[socket.request.session.class].currentStep = 0
-			}
-
-			cpUpdate()
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	// Deletes help ticket
-	socket.on('deleteTicket', (student) => {
-		try {
-			logger.log('info', `[deleteTicket] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[deleteTicket] student=(${student})`)
-
-			classInformation[socket.request.session.class].students[student].help = false
-
-			logger.log('verbose', `[deleteTicket] user=(${JSON.stringify(classInformation[socket.request.session.class].students[student])})`)
-
-			cpUpdate()
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	// Changes the class mode
-	socket.on('modechange', (mode) => {
-		try {
-			logger.log('info', `[modechange] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[modechange] mode=(${mode})`)
-
-			classInformation[socket.request.session.class].mode = mode
-
-			logger.log('verbose', `[modechange] classData=(${classInformation[socket.request.session.class]})`)
-
-			modeUpdate()
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	socket.on('pluginUpdate', () => {
-		logger.log('info', `[pluginUpdate] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-
-		pluginUpdate()
-	})
-
-	socket.on('changePlugin', (id, name, url) => {
-		try {
-			logger.log('info', `[changePlugin] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[changePlugin] id=(${id}) name=(${name}) url=(${url})`)
-
-			if (name) {
-				db.run(
-					'UPDATE plugins set name=? WHERE id=?',
-					[name, id],
-					(err) => {
-						if (err) logger.log('error', err)
-						else pluginUpdate()
-					}
-				)
-			} else if (url) {
-				db.run('UPDATE plugins set url=? WHERE id=?', [url, id], (err) => {
-					if (err) logger.log('error', err)
-					else pluginUpdate()
-				})
-			} else logger.log('critical', 'changePlugin called without name or url')
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	socket.on('addPlugin', (name, url) => {
-		try {
-			logger.log('info', `[addPlugin] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[addPlugin] name=(${name}) url=(${url})`)
-
-			db.get(
-				'SELECT * FROM classroom WHERE key=?',
-				[socket.request.session.class],
-				(err, classData) => {
-					try {
-						if (err) throw err
-
-						db.run(
-							'INSERT INTO plugins(name, url, classId) VALUES(?, ?, ?)',
-							[name, url, classData.id]
-						)
-						pluginUpdate()
-					} catch (err) {
-						logger.log('error', err.stack)
-					}
-				}
-			)
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	socket.on('removePlugin', (id) => {
-		try {
-			logger.log('info', `[removePlugin] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[removePlugin] id=(${id})`)
-
-			db.run('DELETE FROM plugins WHERE id=?', [id])
-			pluginUpdate()
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	socket.on('getOwnedClasses', (username) => {
-		logger.log('info', `[getOwnedClasses] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-		logger.log('info', `[getOwnedClasses] username=(${username})`)
-
-		getOwnedClasses(username)
-	})
-
-	// sends the class code of the class a user is in
-	socket.on('getUserClass', ({ username, api }) => {
-		try {
-			logger.log('info', `[getUserClass] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[getUserClass] username=(${username}) api=(${api})`)
-
-			if (api) {
-				db.get('SELECT * FROM users WHERE API=?', [api], (err, userData) => {
-					try {
-						if (err) throw err
-						if (!userData) {
-							socket.emit('getUserClass', { error: 'not a valid API Key' })
-							return
-						}
-
-						let classCode = getUserClass(userData.username)
-
-						if (classCode instanceof Error) throw classCode
-
-						if (!classCode) socket.emit('getUserClass', { error: 'user is not logged in' })
-						else if (classCode == 'noClass') socket.emit('getUserClass', { error: 'user is not in a class' })
-						else socket.emit('getUserClass', className)
-					} catch (err) {
-						logger.log('error', err.stack)
-						socket.emit('getUserClass', { error: 'There was a server error try again.' })
-					}
-				})
-			} else if (username) {
-				let classCode = getUserClass(username)
-
-				if (classCode instanceof Error) throw classCode
-
-				if (!classCode) socket.emit('getUserClass', { error: 'user is not logged in' })
-				else if (classCode == 'noClass') socket.emit('getUserClass', { error: 'user is not in a class' })
-				else socket.emit('getUserClass', className)
-			} else socket.emit('getUserClass', { error: 'missing username or api key' })
-		} catch (err) {
-			logger.log('error', err.stack)
-			socket.emit('getUserClass', { error: 'There was a server error try again.' })
-		}
-	})
-
-	socket.on('classBannedUsersUpdate', () => {
-		classBannedUsersUpdate()
-	})
-
-	socket.on('classBanUser', (user) => {
-		try {
-			logger.log('info', `[ban] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[ban] user=(${user})`)
-
-			let classCode = socket.request.session.class
-			logger.log('info', `[ban] classCode=(${classCode})`)
-
-			if (!classCode || classCode == 'noClass') {
-				logger.log('info', '[ban] The user is not in a class.')
-				socket.emit('message', 'You are not in a class')
-				return
-			}
-
-			if (!user) {
-				logger.log('critical', '[ban] no username provided.')
-				socket.emit('message', 'No username provided. (Please contact the programmer)')
-				return
-			}
-
-			db.run('UPDATE classusers SET permissions = 0 WHERE classId = (SELECT id FROM classroom WHERE key=?) AND studentId = (SELECT id FROM users WHERE username=?)', [
-				socket.request.session.class,
-				user
-			], (err) => {
-				try {
-					if (err) throw err
-
-					if (classInformation[socket.request.session.class].students[user])
-						classInformation[socket.request.session.class].students[user].classPermissions = 0
-
-					classKickUser(user)
-					advancedEmitToClass('leaveSound', classCode, { api: true })
-					classBannedUsersUpdate()
-					cpUpdate()
-					socket.emit('message', `Banned ${user}`)
-				} catch (err) {
-					logger.log('error', err.stack)
-					socket.emit('message', 'There was a server error try again.')
-				}
-			})
-		} catch (err) {
-			logger.log('error', err.stack)
-			socket.emit('message', 'There was a server error try again.')
-		}
-	})
-
-	socket.on('classUnbanUser', (user) => {
-		try {
-			logger.log('info', `[unban] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[unban] user=(${user})`)
-
-			let classCode = socket.request.session.class
-			logger.log('info', `[unban] classCode=(${classCode})`)
-
-			if (!classCode || classCode == 'noClass') {
-				logger.log('info', '[unban] The user is not in a class.')
-				socket.emit('message', 'You are not in a class')
-				return
-			}
-
-			if (!user) {
-				logger.log('critical', '[unban] no username provided.')
-				socket.emit('message', 'No username provided. (Please contact the programmer)')
-				return
-			}
-
-			db.run('UPDATE classusers SET permissions = 1 WHERE classId = (SELECT id FROM classroom WHERE key=?) AND studentId = (SELECT id FROM users WHERE username=?)', [
-				socket.request.session.class,
-				user
-			], (err) => {
-				try {
-					if (err) throw err
-
-					if (classInformation[socket.request.session.class].students[user])
-						classInformation[socket.request.session.class].students[user].permissions = 1
-
-					classBannedUsersUpdate()
-					socket.emit('message', `Unbanned ${user}`)
-				} catch (err) {
-					logger.log('error', err.stack)
-					socket.emit('message', 'There was a server error try again.')
-				}
-			})
-		} catch (err) {
-			logger.log('error', err.stack)
-			socket.emit('message', 'There was a server error try again.')
-		}
-	})
-
-	socket.on('setClassPermissionSetting', (permission, level) => {
-		try {
-			logger.log('info', `[setClassPermissionSetting] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[setClassPermissionSetting] permission=(${permission}) level=(${level})`)
-
-			let classCode = socket.request.session.class
-			classInformation[classCode].permissions[permission] = level
-			db.run('UPDATE classroom SET permissions=? WHERE id=?', [JSON.stringify(classInformation[classCode].permissions), classInformation[classCode].id], (err) => {
-				try {
-					if (err) throw err
-
-					logger.log('info', `[setClassPermissionSetting] ${permission} set to ${level}`)
-					cpUpdate()
-				} catch (err) {
-					logger.log('error', err.stack)
-				}
-			})
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	socket.on('deleteUser', async (userId) => {
-		try {
-			logger.log('info', `[deleteUser] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[deleteUser] userId=(${userId})`)
-
-			const user = await new Promise((resolve, reject) => {
-				db.get('SELECT * FROM users WHERE id=?', userId, (err, user) => {
-					if (err) reject(err)
-					resolve(user)
-				})
-			})
-			if (!user) {
-				socket.emit('message', 'User not found')
-				return
-			}
-
-			if (userSockets[user.username])
-				logout(userSockets[user.username])
-
-			try {
-				await runQuery('BEGIN TRANSACTION')
-
-				await Promise.all([
-					runQuery('DELETE FROM users WHERE id=?', userId),
-					runQuery('DELETE FROM classusers WHERE studentId=?', userId),
-					runQuery('DELETE FROM shared_polls WHERE userId=?', userId),
-				])
-
-				await deleteCustomPolls(userId)
-				await deleteClassrooms(userId)
-
-				await runQuery('COMMIT')
-				await managerUpdate()
-				socket.emit('message', 'User deleted successfully')
-			} catch (err) {
-				await runQuery('ROLLBACK')
-				throw err
-			}
-		} catch (err) {
-			logger.log('error', err.stack)
-			socket.emit('message', 'There was a server error try again.')
-		}
-	})
-
-	socket.on('ipUpdate', () => {
-		ipUpdate(null, socket.request.session.username)
-	})
-
-	socket.on('changeIp', (type, id, ip) => {
-		try {
-			logger.log('info', `[changeIp] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[changeIp] type=(${type}) id=(${id}) ip=(${ip})`)
-
-			if (type != 'whitelist' && type != 'blacklist') {
-				logger.log('critical', 'invalid type')
-				socket.emit('message', 'Invalid Ip type')
-				return
-			}
-
-			db.get(`SELECT * FROM ip_${type} WHERE id=?`, id, (err, dbIp) => {
-				if (err) {
-					logger.log('error', err.stack)
-					socket.emit('message', 'There was a server error try again.')
-					return
-				}
-
-				if (!dbIp) {
-					socket.emit('message', 'Ip not found')
-					return
-				}
-
-
-				db.run(`UPDATE ip_${type} set ip=? WHERE id=?`, [ip, id], (err) => {
-					if (err) logger.log('error', err)
-					else {
-						if (type == 'whitelist') whitelistedIps[dbIp.id].ip = ip
-						else if (type == 'blacklist') blacklistedIps[dbIp.id].ip = ip
-
-
-						reloadPageByIp(type == 'whitelist', ip)
-						reloadPageByIp(type == 'whitelist', dbIp.ip)
-						ipUpdate(type)
-					}
-				})
-			})
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	socket.on('addIp', (type, ip) => {
-		logger.log('info', `[addIp] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-		logger.log('info', `[addIp] type=(${type}) ip=(${ip})`)
-
-		if (type != 'whitelist' && type != 'blacklist') {
-			logger.log('critical', 'invalid type')
-			socket.emit('message', 'Invalid Ip type')
-			return
-		}
-
-		db.get(`SELECT * FROM ip_${type} WHERE ip=?`, ip, (err, dbIp) => {
-			if (err) {
-				logger.log('error', err.stack)
-				socket.emit('message', 'There was a server error try again.')
-				return
-			}
-
-			if (dbIp) {
-				socket.emit('message', `IP already in ${type}`)
-				return
-			}
-
-			db.run(`INSERT INTO ip_${type} (ip) VALUES(?)`, [ip], (err) => {
-				if (err) {
-					logger.log('error', err.stack)
-					socket.emit('message', 'There was a server error try again.')
-					return
-				}
-
-				db.get(`SELECT * FROM ip_${type} WHERE ip=?`, ip, (err, dbIp) => {
-					if (err) {
-						logger.log('error', err.stack)
-						socket.emit('message', 'There was a server error try again.')
-						return
-					}
-
-					if (type == 'whitelist') whitelistedIps[dbIp.id] = dbIp
-					else if (type == 'blacklist') blacklistedIps[dbIp.id] = dbIp
-
-					reloadPageByIp(type != 'whitelist', ip)
-					ipUpdate(type)
-					socket.emit('message', `IP added to ${type}`)
-				})
-			})
-		})
-	})
-
-	socket.on('removeIp', (type, id) => {
-		try {
-			logger.log('info', `[removeIp] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[removeIp] type=(${type}) id=(${id})`)
-
-			if (type != 'whitelist' && type != 'blacklist') {
-				logger.log('critical', 'invalid type')
-				socket.emit('message', 'Invalid Ip type')
-				return
-			}
-
-			db.get(`SELECT * FROM ip_${type} WHERE id=?`, id, (err, dbIp) => {
-				if (err) {
-					logger.log('error', err)
-					socket.emit('message', 'There was a server error try again.')
-					return
-				}
-
-				if (!dbIp) {
-					socket.emit('message', 'Ip not found')
-					return
-				}
-
-				db.run(`DELETE FROM ip_${type} WHERE id=?`, [id], (err) => {
-					if (err) {
-						logger.log('error', err)
-						socket.emit('message', 'There was a server error try again.')
-						return
-					}
-
-					reloadPageByIp(type != 'whitelist', dbIp.ip)
-					if (type == 'whitelist') delete whitelistedIps[id]
-					else if (type == 'blacklist') delete blacklistedIps[id]
-					ipUpdate(type)
-				})
-			})
-		} catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	socket.on('toggleIpList', (type) => {
-		logger.log('info', `[toggleIpList] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-		logger.log('info', `[toggleIpList] type=(${type})`)
-
-		if (type != 'whitelist' && type != 'blacklist') {
-			logger.log('critical', 'invalid type')
-			socket.emit('message', 'Invalid Ip type')
-			return
-		}
-
-		settings[`${type}Active`] = !settings[`${type}Active`]
-		fs.writeFileSync('./settings.json', JSON.stringify(settings))
-
-		let ipList
-		if (type == 'whitelist') ipList = whitelistedIps
-		else if (type == 'blacklist') ipList = blacklistedIps
-
-		for (let ip of Object.values(ipList)) {
-			reloadPageByIp(type != 'whitelist', ip.ip)
-		}
-		ipUpdate(type)
-	})
-
-	socket.on('saveTags', (studentId, tags, username) => {
-		//Save the tags to the students tag element in their object
-		//Then save their tags to the database
-		try {
-			logger.log('info', `[saveTags] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
-			logger.log('info', `[saveTags] studentId=(${studentId}) tags=(${JSON.stringify(tags)})`)
-			classInformation[socket.request.session.class].students[username].tags = tags.toString()
-			db.get('SELECT tags FROM users WHERE id=?', [studentId], (err, row) => {
-				if (err) {
-					logger.log('error', err)
-					socket.emit('message', 'There was a server error try again.')
-					return
-				}
-				if (row) {
-					// Row exists, update it
-					db.run('UPDATE users SET tags=? WHERE id=?', [tags.toString(), studentId], (err) => {
-						if (err) {
-							logger.log('error', err)
-							socket.emit('message', 'There was a server error try again.')
-							return
-						}
-					});
-				} else {
-					socket.send('message', 'User not found')
-				}
-			});
-		}
-		catch (err) {
-			logger.log('error', err.stack)
-		}
-	})
-
-	socket.on('newTag', (tagName) => {
-		//Add a new tag to the database
-		try {
-			if (tagName == '') return;
-			classInformation[socket.request.session.class].tagNames.push(tagName);
-			var newTotalTags = "";
-			for (let i = 0; i < classInformation[socket.request.session.class].tagNames.length; i++) {
-				newTotalTags += classInformation[socket.request.session.class].tagNames[i] + ", ";
-			};
-			newTotalTags = newTotalTags.split(", ");
-			newTotalTags.pop();
-			db.get('SELECT * FROM classroom WHERE name = ?', [classInformation[socket.request.session.class].className], (err, row) => {
-				if (err) {
-					logger.log(err.stack);
-				}
-				if (row) {
-					db.run('UPDATE classroom SET tags = ? WHERE name = ?', [newTotalTags.toString(), classInformation[socket.request.session.class].className], (err) => {
-						if (err) {
-							logger.log(err.stack);
-						};
-					});
-				} else {
-					socket.send('message', 'Class not found')
-				};
-			});
-		}
-		catch (err) {
-			logger.log('error', err.stack);
-		}
-	})
-
-	socket.on('removeTag', (tagName) => {
-		try {
-			//Find the tagName in the array of tagnames from the database
-			//If the tagname is not there, socket.send('message', 'Tag not found') and return
-			//If the tagname is there, remove it from the array and update the database
-			var index = classInformation[socket.request.session.class].tagNames.indexOf(tagName);
-			if (index > -1) {
-				classInformation[socket.request.session.class].tagNames.splice(index, 1);
-			} else {
-				socket.send('message', 'Tag not found')
-				return;
-			}
-			//Now remove all instances of the tag from the students' tags
-			for (let student of Object.values(classInformation[socket.request.session.class].students)) {
-				if (student.classPermissions == 0 || student.classPermissions >= 5) continue;
-				var studentTags = student.tags.split(",");
-				var studentIndex = studentTags.indexOf(tagName);
-				if (studentIndex > -1) {
-					studentTags.splice(studentIndex, 1);
-				}
-				student.tags = studentTags.toString();
-				db.get('SELECT * FROM users WHERE username = ?', [student.username], (err, row) => {
-					if (err) {
-						logger.log(err.stack);
-					}
-					if (row) {
-						db.run('UPDATE users SET tags = ? WHERE username = ?', [studentTags.toString(), student.username], (err) => {
-							if (err) {
-								logger.log(err.stack);
-							};
-						});
-					} else {
-						socket.send('message', 'User not found')
-					};
-				});
-				db.get('SELECT tags FROM classroom WHERE name = ?', [classInformation[socket.request.session.class].className], (err, row) => {
-					if (err) {
-						logger.log(err.stack);
-					}
-					//Set the tags in the database to a variable
-					//Remove the tag from the variable
-					//Update the database with the new variable
-					if (row) {
-						var newTotalTags = row.tags;
-						newTotalTags = newTotalTags.split(",");
-						var tagIndex = newTotalTags.indexOf(tagName);
-						if (tagIndex > -1) {
-							newTotalTags.splice(tagIndex, 1);
-						}
-						db.run('UPDATE classroom SET tags = ? WHERE name = ?', [newTotalTags.toString(), classInformation[socket.request.session.class].className], (err) => {
-							if (err) {
-								logger.log(err.stack);
-							};
-						});
-					} else {
-						socket.send('message', 'Class not found')
-					};
-				})
-			};
-		}
-		catch (err) {
-			logger.log('error', err.stack);
-		}
-	});
-
-	socket.on("approvePasswordChange", (changeApproval, username, newPassword) => {
-		try {
-			if (changeApproval) {
-				let passwordCrypt = encrypt(newPassword);
-				let passwordCryptString = JSON.stringify(passwordCrypt);
-				db.run("UPDATE users SET password = ? WHERE username = ?", [passwordCryptString, username], (err) => {
-					if (err) {
-						logger.log("error", err.stack);
-					};
-				});
-			};
-		} catch (err) {
-			logger.log("error", err.stack);
-		};
-	});
-
-	socket.on("classPoll", (poll) => {
-		try {
-			let userId = socket.request.session.userId
-			db.get('SELECT seq AS nextPollId from sqlite_sequence WHERE name = "custom_polls"', (err, nextPollId) => {
-				try {
-					if (err) throw err
-					if (!nextPollId) logger.log('critical', '[savePoll] nextPollId not found')
-
-					nextPollId = nextPollId.nextPollId + 1
-
-					db.run('INSERT INTO custom_polls (owner, name, prompt, answers, textRes, blind, weight, public) VALUES (?, ?, ?, ?, ?, ?, ?, ?)', [
-						userId,
-						poll.name,
-						poll.prompt,
-						JSON.stringify(poll.answers),
-						poll.textRes,
-						poll.blind,
-						poll.weight,
-						poll.public
-					], (err) => {
-						try {
-							if (err) throw err
-
-							classInformation[socket.request.session.class].students[socket.request.session.username].ownedPolls.push(nextPollId)
-							socket.emit('message', 'Poll saved successfully!')
-							customPollUpdate(socket.request.session.username)
-							socket.emit("classPollSave", nextPollId);
-						} catch (err) {
-							logger.log('error', err.stack);
-						}
-					})
-				} catch (err) {
-					logger.log('error', err.stack);
-				}
-			})
-		} catch (err) {
-			logger.log("error", err.stack);
-		}
-	})
-
-	socket.on('vbTimer', () => {
-		let classData = classInformation[socket.request.session.class];
-		let username = socket.request.session.username
-
-		advancedEmitToClass('vbTimer', socket.request.session.class, {
-			classPermissions: CLASS_SOCKET_PERMISSIONS.vbTimer,
-			username
-		}, classData.timer);
-	})
-
-	socket.on("timer", (startTime, active, sound) => {
-		//This handles the server side timer
-		try {
-			let classData = classInformation[socket.request.session.class];
-
-			startTime = Math.round(startTime * 60)
-
-			classData.timer.startTime = startTime
-			classData.timer.timeLeft = startTime + 1
-			classData.timer.active = active
-			classData.timer.sound = sound
-
-			cpUpdate(socket.request.session.class)
-			if (active) {
-				//run the function once instantly
-				timer(sound, active)
-				//save a clock in the class data, that way it saves when the page is refreshed
-				runningTimers[socket.request.session.class] = setInterval(() => timer(sound, active), 1000);
-			} else {
-				//if the timer is not active, clear the interval
-				clearInterval(runningTimers[socket.request.session.class]);
-				runningTimers[socket.request.session.class] = null;
-
-				timer(sound, active)
-			}
-		} catch (err) {
-			logger.log("error", err.stack);
-		}
-	})
-
-	socket.on("timerOn", () => {
-		socket.emit("timerOn", classInformation[socket.request.session.class].timer.active);
-	})
-
-})
-
 
 http.listen(420, async () => {
 	whitelistedIps = await getIpAccess('whitelist')
