+++ crypto.js
@@ -3,11 +3,9 @@ const crypto = require('crypto');
 const algorithm = 'aes-256-ctr';
 const secretKey = 'vOVH6sdmpNWjRRIqCc7rdxs01lwHzfr3';
 
-const encrypt = (text) => {
+function encrypt(text) {
     const iv = crypto.randomBytes(16);
-
     const cipher = crypto.createCipheriv(algorithm, secretKey, iv);
-
     const encrypted = Buffer.concat([cipher.update(text), cipher.final()]);
 
     return {
@@ -16,12 +14,11 @@ const encrypt = (text) => {
     };
 };
 
-const decrypt = (hash) => {
+function decrypt(hash) {
     const decipher = crypto.createDecipheriv(algorithm, secretKey, Buffer.from(hash.iv, 'hex'));
+    const decrypted = Buffer.concat([decipher.update(Buffer.from(hash.content, 'hex')), decipher.final()]);
 
-    const decrpyted = Buffer.concat([decipher.update(Buffer.from(hash.content, 'hex')), decipher.final()]);
-
-    return decrpyted.toString();
+    return decrypted.toString();
 };
 
 module.exports = {
