+++ routes/index.js
@@ -3,12 +3,16 @@
 // This allows it to check if the user is logged in along with the home page
 // It also allows for redirection to any other page if needed
 
-export default {
-    run() {
+const { isAuthenticated } = require("../modules/authentication")
+const { classInformation } = require("../modules/class")
+const logger = require("../modules/logger")
+
+module.exports = {
+    run(app) {
         app.get('/', isAuthenticated, (req, res) => {
             try {
                 logger.log('info', `[get /] ip=(${req.ip}) session=(${JSON.stringify(req.session)})`)
-                if (cD[req.session.class].students[req.session.username].classPermissions >= TEACHER_PERMISSIONS) {
+                if (classInformation[req.session.class].students[req.session.username].classPermissions >= TEACHER_PERMISSIONS) {
                     res.redirect('/controlPanel')
                 } else {
                     res.redirect('/student')
